% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{as.ts.mcmc}
\alias{as.ts.mcmc}
\title{Copied from coda package source, to fulfill
 autocorr.plot requirement}
\usage{
\method{as.ts}{mcmc}(x, ...)
}
\arguments{
\item{x}{an mcmc object}

\item{...}{unused arguments for compatibility with generic as.ts}
}
\description{
Copied from coda package source, to fulfill
 autocorr.plot requirement
}
