% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{as.tsmcmc}
\alias{as.tsmcmc}
\title{Copied from coda package source, to fulfill
 autocorr.plot requirement}
\usage{
as.tsmcmc(x, ...)
}
\arguments{
\item{x}{an mcmc object}

\item{...}{unused arguments for compatibility with [coda::as.ts.mcmc()]}
}
\description{
Copied from coda package source, to fulfill
 autocorr.plot requirement
}
