% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorr-plot.R
\name{autocorr_plot}
\alias{autocorr_plot}
\title{An extended version of \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}
\usage{
autocorr_plot(x, lag_max, auto_layout = TRUE, ...)
}
\arguments{
\item{x}{See \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}

\item{lag_max}{\code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}

\item{auto_layout}{\code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}}

\item{...}{Arguments passed to \code{\link[base:plot]{base::plot()}}}
}
\value{
Nothing
}
\description{
An extended version of \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}}
}
\details{
This code was copied exactly from \code{\link[coda:autocorr.plot]{coda::autocorr.plot()}} and modified to pass
arguments to \code{\link[base:plot]{base::plot()}} and to remove the \code{ask} argument.
}
