% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{biomass.plot.mpd}
\alias{biomass.plot.mpd}
\title{Plot the MPD biomass time series along with the associated retrospectives.}
\usage{
biomass.plot.mpd(
  model,
  depl = FALSE,
  xlim = NA,
  offset = 0.7,
  leg = TRUE,
  a_trans = 0.75,
  abund = TRUE,
  df,
  translate = FALSE
)
}
\arguments{
\item{model}{an iscam model object}

\item{depl}{logical; if TRUE, plot the depletion line(s)}

\item{xlim}{numeric vector of length 2; trim the x-axis (default NA)}

\item{offset}{horizontal offset for B0 points in depletion plot}

\item{leg}{show the legend? Logical}

\item{a_trans}{transparency for lines and points}

\item{abund}{logical; plot scaled abundance (points)}

\item{df}{Data frame of the survey estimates, as constructed by [get_surv_ind()]}

\item{translate}{logical; translate labels}
}
\value{
a ggplot object
}
\description{
Biomass can be absolute, or relative (i.e., depletion). Also plot estimated
unfished spawning biomass as points.
}
