% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{build_herring_rdata_files}
\alias{build_herring_rdata_files}
\title{Create RData files for models found in the directories given by `dirs`}
\usage{
build_herring_rdata_files(dirs, ...)
}
\arguments{
\item{dirs}{List of directories to build Rdata files for}

\item{...}{arguments passed to [create.rdata.file()]}
}
\description{
Create RData files for models found in the directories given by `dirs`
}
