% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{c.model.list}
\alias{c.model.list}
\title{Extract the model class objects from the list of model lists,
and merge them into a single model list containing all the model
class objects}
\usage{
\method{c}{model.list}(...)
}
\arguments{
\item{...}{one or more lists of iscam model objects}
}
\value{
an iscam model object list
}
\description{
Extract the model class objects from the list of model lists,
and merge them into a single model list containing all the model
class objects
}
