% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.ahat}
\alias{calc.ahat}
\title{Calculate the estimated numbers-at-age for an iscam MCMC model}
\usage{
calc.ahat(model)
}
\arguments{
\item{model}{An iscam model object}
}
\value{
A list of data frames, one for each gear
}
\description{
Calculate the estimated numbers-at-age for an iscam MCMC model
}
