% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.mcmc}
\alias{calc.mcmc}
\title{Perform some quantile calculations on the MCMC posteriors}
\usage{
calc.mcmc(model, burnin = 1000, thin = 1, lower = 0.025,
  upper = 0.975, load.proj = TRUE, which.stock = NULL,
  which.model = NULL, fixed.cutoffs)
}
\arguments{
\item{model}{An iscam model object}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{thin}{The thinning to apply to the MCMC posterior samples}

\item{lower}{Lower quantile value to apply to MCMC samples}

\item{upper}{Upper quantile value to apply to MCMC samples}

\item{load.proj}{Load the projections from the MCMC and do the calculations
to construct the decision tables}

\item{which.stock}{1-5 for the five herring stocks: 1=HG, 2=PRD, 3=CC,
4=SOG, 5=WCVI}

\item{which.model}{1 = AM1 or 2 = AM2 for herring}

\item{fixed.cutoffs}{A vector of catch cutoffs to use in decision tables}
}
\value{
A list of each parameter for which quantiles were calculated
}
\description{
Perform some quantile calculations on the MCMC posteriors
}
