% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.mpd.logs}
\alias{calc.mpd.logs}
\title{Calculate logs for several parameters using MPD}
\usage{
calc.mpd.logs(
  mpd,
  log.params = c("^ro$", "^m$", "^m1$", "^m2$", "^rbar$", "^rinit$", "^q$")
)
}
\arguments{
\item{mpd}{A list of MPD outputs}

\item{log.params}{The names of the parameters to log}
}
\value{
a list (mpd with some new elements appended, the log-applied elements)
}
\description{
Calculate logs for several parameters using MPD
}
\details{
The new list elements will have the same names but have 'log_' appended
}
