% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.probabilities}
\alias{calc.probabilities}
\title{Extract and calculate probabilities from the iscam projection model}
\usage{
calc.probabilities(model, burnin, thin, which.stock = NULL,
  which.model = NULL, fixed.cutoffs)
}
\arguments{
\item{model}{An iscam model object}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{thin}{The thinning to apply to the MCMC posterior samples}

\item{which.stock}{1-5 for the five herring stocks: 1=HG, 2=PRD, 3=CC,
4=SOG, 5=WCVI}

\item{which.model}{1 = AM1 or 2 = AM2 for herring}

\item{fixed.cutoffs}{A vector of catch cutoffs to use in decision tables}
}
\value{
A data frame which has its names formatted for latex
}
\description{
Extract and calculate probabilities from the iscam projection model
}
\details{
Extract and calculate probabilities from the projection model.
  Used for decision tables in the document (see make.decision.table())
  in tables-decisions.r
}
