% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc.sig.tau}
\alias{calc.sig.tau}
\title{Calculation of sigma and tau from rho and vartheta}
\usage{
calc.sig.tau(rho, vartheta)
}
\arguments{
\item{rho}{parameter rho from iscam model}

\item{vartheta}{parameter vartheta from iscam model}
}
\value{
a list of length 2, the calculated tau and sigma parameters
}
\description{
Calculation of sigma and tau from rho and vartheta
}
