% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_age_quants}
\alias{calc_age_quants}
\title{Calculate age fits / age residuals quantiles}
\usage{
calc_age_quants(lst, probs)
}
\description{
@param lst A list as output by [load_agefit()]
 of either MCMC age fits or residuals
 @param probs A vector of three values for quantiles calculations
 @export
}
