% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_beta_mean_cv}
\alias{calc_beta_mean_cv}
\title{Calculate the mean and CV for a Beta distribution given the alpha and
beta shape parameters}
\usage{
calc_beta_mean_cv(alpha, beta)
}
\arguments{
\item{alpha}{Shape parameter 1}

\item{beta}{Shape parameter 2}
}
\value{
A vector of two values, mean and CV
}
\description{
Calculate the mean and CV for a Beta distribution given the alpha and
beta shape parameters
}
