% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_beta_params}
\alias{calc_beta_params}
\title{Calculate the shape parameters (alpha and beta) for a Beta distribution
given the mean and standard deviation}
\usage{
calc_beta_params(mu, cv)
}
\arguments{
\item{mu}{Mean of the distribution}

\item{sd}{Standard deviation of the distribution}
}
\value{
A vector of two values, alpha and beta
}
\description{
Calculate the shape parameters (alpha and beta) for a Beta distribution
given the mean and standard deviation
}
