% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_longer_quants}
\alias{calc_longer_quants}
\title{Calculate age fits / age residuals and selectivity estimates quantiles}
\usage{
calc_longer_quants(lst, type = c("age", "sel"), probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{lst}{A list as output by \code{\link[=load_longer]{load_longer()}}
of either MCMC age fits, age residuals, or selectivity estimates}

\item{type}{The type of quantiles to calculate}

\item{probs}{A vector of three values for quantiles calculations}
}
\description{
Calculate age fits / age residuals and selectivity estimates quantiles
}
