% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc_mcmc}
\alias{calc_mcmc}
\title{Perform some quantile calculations on the MCMC posteriors}
\usage{
calc_mcmc(model, probs = c(0.025, 0.5, 0.975), load_proj = TRUE, ...)
}
\arguments{
\item{model}{An iSCAM model object as created in [load_iscam_files()]}

\item{probs}{The probabilities to use for `quantile()` calculations}

\item{load_proj}{Load the projections from the MCMC and do the calculations}

\item{...}{arguments to pass to [calc_probabilities()]
to construct the decision tables}
}
\value{
A list of each parameter for which quantiles were calculated
}
\description{
Perform some quantile calculations on the MCMC posteriors
}
