% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc_mpd_logs}
\alias{calc_mpd_logs}
\title{Calculate logs for several parameters using MPD}
\usage{
calc_mpd_logs(mpd, log_params = c("^ro$", "^m$", "^rbar$", "^rinit$", "^q$"))
}
\arguments{
\item{mpd}{A list of MPD outputs}

\item{log_params}{Patterns to match for parameter names that need to be logged}
}
\value{
a list (mpd with some new elements appended, the log-applied elements)
}
\description{
Calculate logs for several parameters using MPD
}
\details{
The new list elements will have the same names but have 'log_' appended
}
