% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-probs-biomass.R
\name{calc_probs_biomass}
\alias{calc_probs_biomass}
\title{Calculate the probabilities that biomasses are under or over given
reference points}
\usage{
calc_probs_biomass(d, refpt = 0.2, prob_gt = TRUE, format = c("latex", "html"))
}
\arguments{
\item{d}{A data frame containing a column named \code{catch} and other columns
names as years of depletion}

\item{refpt}{The depletion value that is checked against in the data frame
\code{d}}

\item{prob_gt}{Stand for 'probability greater than'. If \code{TRUE},}

\item{format}{If 'latex' modify column headers with latex code, if
'html' modify column headers with html code}
}
\description{
Calculate the probabilities that biomasses are under or over given
reference points or that they decline or increase from one year to the
next
}
