% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-proj-probs.R
\name{calc_proj_probs}
\alias{calc_proj_probs}
\title{Create decision table from iSCAM MCMC model output}
\usage{
calc_proj_probs(model, fixed_cutoffs)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{fixed_cutoffs}{A vector of catch cutoffs to use in decision tables}
}
\value{
A data frame which has its names formatted for latex
}
\description{
Extract and calculate probabilities from the iscam projection model
}
\details{
Extract and calculate probabilities from the projection model.
Used for decision tables in the document (see make.decision.table())
in tables-decisions.r
}
