% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_rho_vartheta}
\alias{calc_rho_vartheta}
\title{Calculation of rho and vartheta from sigma and tau}
\usage{
calc_rho_vartheta(sig, tau)
}
\arguments{
\item{sig}{Parameter sigma}

\item{tau}{Parameter tau}
}
\value{
A vector of two values, rho and vartheta
}
\description{
Calculation of rho and vartheta from sigma and tau
}
\seealso{
calc_sig_tau
}
