% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_sig_tau}
\alias{calc_sig_tau}
\title{Calculation of sigma and tau from rho and vartheta}
\usage{
calc_sig_tau(rho, vartheta)
}
\arguments{
\item{rho}{Parameter rho from iscam model}

\item{vartheta}{Parameter vartheta from iscam model
(actually vartheta squared in the iSCAM equation T5.2)}
}
\value{
A vector of two values, sigma and tau
}
\description{
Calculation of sigma and tau from rho and vartheta
}
\details{
Total variance is given by ϑ (\code{vartheta}) and the proportion of the total
variance that is associated with observation errors is given by ρ (\code{rho}),
the variance is partitioned into observation errors (σ^2) (\code{sigma}^2) and
process errors (τ^2) (\code{tau}^2).
}
