% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CalcXYLims}
\alias{CalcXYLims}
\title{Set the space between the points and the plot border manually as a specified
percentage of the range. Return a list with range for x and y}
\usage{
CalcXYLims(x, y, pX = 0.01, pY = 0.01)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{pX}{pX}

\item{pY}{pY}
}
\description{
Set the space between the points and the plot border manually as a specified
percentage of the range. Return a list with range for x and y
}
