% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ClipExtent}
\alias{ClipExtent}
\title{Clip to the extent of a supplied sp object}
\usage{
ClipExtent(dat, spObj, bufDist = NA, silent = FALSE)
}
\arguments{
\item{dat}{dat}

\item{spObj}{spObj}

\item{bufDist}{bufDist}

\item{silent}{silent}
}
\description{
Clip to the extent of a supplied sp object
}
