% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colr}
\alias{colr}
\title{A simple function for creating transparent colors}
\usage{
colr(col.pal = 1, a = 1)
}
\arguments{
\item{col.pal}{Color value}

\item{a}{alpha value}
}
\description{
A simple function for creating transparent colors
}
