% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert-prob-cols-language.R
\name{convert_prob_cols_language}
\alias{convert_prob_cols_language}
\title{Convert any columns in a data frame that appear as 22.5\% (for example)
to 22,5\% which is how it appears in French}
\usage{
convert_prob_cols_language(d)
}
\arguments{
\item{d}{A data frame}
}
\value{
\code{d}, but with column names modified as described in the function
title
}
\description{
Convert any columns in a data frame that appear as 22.5\% (for example)
to 22,5\% which is how it appears in French
}
