% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ConvertLineEndings}
\alias{ConvertLineEndings}
\title{Convert line endings to Linux}
\usage{
ConvertLineEndings(infile)
}
\arguments{
\item{infile}{infile}
}
\description{
Convert line endings to Linux
}
