% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountConsecutive}
\alias{CountConsecutive}
\title{Function to add a new column indicating the number of consecutive values}
\usage{
CountConsecutive(vec)
}
\arguments{
\item{vec}{vector}
}
\description{
Function to add a new column indicating the number of consecutive values
}
