% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{create.rdata.file.retro}
\alias{create.rdata.file.retro}
\title{Create a .RData file to hold the restrospective model's data and outputs}
\usage{
create.rdata.file.retro(model.dir, ovwrt.rdata = FALSE,
  load.proj = TRUE, low = 0.025, high = 0.975, burnin = 1000,
  thin = 1, which.stock = NULL, which.model = NULL,
  verbose = FALSE)
}
\arguments{
\item{model.dir}{Directory name for all models location}

\item{ovwrt.rdata}{overwrite the RData file if it exists? TRUE/FALSE}

\item{load.proj}{Load the projections from the mcmc and do the calculations
to construct the decision tables}

\item{low}{Lower quantile value to apply to MCMC samples}

\item{high}{Higher quantile value to apply to MCMC samples}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{which.stock}{1-5 for the five herring stocks: 1=HG, 2=PRD, 3=CC,
4=SOG, 5=WCVI}

\item{which.model}{1 = AM1 or 2 = AM2 for herring}

\item{verbose}{Say more}
}
\value{
Nothing
}
\description{
Create a .RData file to hold the restrospective model's data and outputs
}
\details{
Create a .RData file to hold the restrospective model's data and outputs.
  If an RData file exists, and overwrite is FALSE, return immediately.
  If no RData file exists, the model will be loaded from outputs into
  an R list and saved as an RData file in the correct directory.
  When this function exits, an RData file will be located in the
  directory given by model.name.
  Assumes the files model-setup.r and utilities.r has been sourced.
}
