% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-file.R
\name{create_rds_file}
\alias{create_rds_file}
\title{Create an rds file to hold the model's data and outputs.}
\usage{
create_rds_file(model_dir = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{model_dir}{Directory name of model to be loaded}

\item{overwrite}{Logical. If TRUE, overwrite the RDS file if it exists}
}
\value{
[base::invisible()]
}
\description{
Create an rds file to hold the model's data and outputs.
}
