% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{delete.dirs}
\alias{delete.dirs}
\title{Delete all directories and files in a directory}
\usage{
delete.dirs(models.dir = model.dir, sub.dir = NULL)
}
\arguments{
\item{models.dir}{Directory name for all models location}

\item{sub.dir}{The subdirectory to delete recursively from}
}
\value{
Nothing
}
\description{
Delete all directories and files in a directory
}
