% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{delete_files_ext}
\alias{delete_files_ext}
\title{Delete all files with a specific extension recursively. By default, .rds files}
\usage{
delete_files_ext(path = NULL, ext = "rds")
}
\arguments{
\item{path}{The directory to start in.}

\item{ext}{The extension of files to include for deletion}
}
\description{
Delete all files with a specific extension recursively. By default, .rds files
}
