% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DeltaPercent}
\alias{DeltaPercent}
\title{Calculate percent change, difference, etc}
\usage{
DeltaPercent(x, nYrs = 1, type)
}
\arguments{
\item{x}{val}

\item{nYrs}{num yrs}

\item{type}{"PctChange" or "PctDiff"}
}
\value{
value
}
\description{
Calculate percent change, difference, etc
}
