% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{dos.pipe.to.log}
\alias{dos.pipe.to.log}
\title{The DOS pipe to log operator string}
\format{
An object of class \code{character} of length 1.
}
\usage{
dos.pipe.to.log
}
\description{
Used to pipe output to a log file when running DOS bat files from  within
R
}
\keyword{datasets}
