% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{draw.envelope}
\alias{draw.envelope}
\title{Draw a time series envelope on a device on which [plot.new()] has already been called}
\usage{
draw.envelope(yrs, quants, col = "black", first, opacity = 75, ...)
}
\arguments{
\item{yrs}{the years to plot}

\item{quants}{a 3-row matrix, where the middle row is the median and the other two are
the lower and upper values for some confidence interval.}

\item{col}{color of the envelope}

\item{first}{boolean. If TRUE, [plot.new()] will be called. If FALSE, [lines()] will be
called.}

\item{opacity}{how opaque the envelope shading is. Percentage value}

\item{...}{other graphical parameters}
}
\description{
Draw a time series envelope on a device on which [plot.new()] has already been called
}
