% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{extract_age_output}
\alias{extract_age_output}
\title{Extract an age structure MPD object from iSCAM output (.rep file)}
\usage{
extract_age_output(model, type = "obs")
}
\arguments{
\item{model}{An iscam model object}

\item{type}{One of 'obs', 'est', 'resid'}
}
\value{
A list of data frames, one for each gear
}
\description{
Extract an age structure MPD object from iSCAM output (.rep file)
}
