% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-area-sex-matrices.R
\name{extract_area_sex_matrices}
\alias{extract_area_sex_matrices}
\title{Extract the given data frame into a list of matrices by iscam 'area'
and 'sex'}
\usage{
extract_area_sex_matrices(data = NULL, prefix = NULL)
}
\arguments{
\item{data}{A data frame read in from one of the MCMC csv output files}

\item{prefix}{See details}
}
\value{
a list (area-sex) of lists (gears) of matrices, one element
per group
}
\description{
Extract the given data frame into a list of matrices by iscam 'area'
and 'sex'
}
\details{
Extract the data frame given (data) by unflattening into a list of matrices
by area-sex and gear. The area-sex number is located in the names of the
columns of the data frame in this format:
"prefix(areasexnum)_gear(gearnum)_year" where (areasexnum) and (gearnum)
are one or more digits and prefix is the string given as an argument
to the function.
}
