% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{fetch_file_names}
\alias{fetch_file_names}
\title{Read the iscam starter file to get the iscam input file names}
\usage{
fetch_file_names(path, filename)
}
\arguments{
\item{path}{Full path to the file}

\item{filename}{Filename}
}
\value{
A list with three names:
  1. Data file name
  2. Control file name
  3. Projection file name
}
\description{
Read the iscam starter file to get the iscam input file names
}
