% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_f_b40.R
\name{find_f_b40}
\alias{find_f_b40}
\title{Find the Reference Removal Rate (RRR) \code{bo_refpt} F_b0 which is the F value
which  will bring the Spawning biomass to \code{bo_refpt} B0 in one year within
the tolerance of \code{tol} thousand tonnes.}
\usage{
find_f_b40(
  model,
  bo_refpt = 0.4,
  init_catch = 1,
  root_dir = "d",
  tol = 0.5,
  max_iter = 10,
  burnin = 1000,
  proj_yrs = 50
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{bo_refpt}{A fractional value of B0 to use as the reference point to
approach in the calculation}

\item{init_catch}{Initial catch value to run the model with}

\item{root_dir}{The root hard drive name, must be either 'c', 'C',
'd', or 'D'}

\item{tol}{The catch tolerance in thousands of tonnes, default 0.5}

\item{max_iter}{The maximum number of iterations to perform. If this is
hit, a warning will be issued and the F value returned will be NA}

\item{burnin}{The number of posterior samples to remove from the beginning
of the list of posteriors}

\item{proj_yrs}{The number of years to project}
}
\value{
A list of two items, the F and U values. Each of these are the length
of the number of fleets
}
\description{
Find the Reference Removal Rate (RRR) \code{bo_refpt} F_b0 which is the F value
which  will bring the Spawning biomass to \code{bo_refpt} B0 in one year within
the tolerance of \code{tol} thousand tonnes.
}
\details{
This is performed by an automated search for the catch level required to
get the F value for the removal rate reference point. The initial catch is
provided and should be based on the value in the decision table which
is close to 0.5 probability that the biomass next year will drop
below \code{bo_refpt} of B0. The initial value is placed as a catch TAC value
in the projection file and the model mceval is run, then the biomass is
compared with the 0.4B0 value and the tolerance is checked. If within the
tolerance, \code{tol}, the F value is extracted and that is the RRR. If it is
not within tolerance \code{tol}, the new catch becomes the current evaluated
catch plus the difference between the estimated biomass and the B0
reference point biomass. This routine is continued until within the
tolerance \code{tol} or the maximum number of iterations has taken place
(\code{max_iter}).

The \code{bash} shell is used to run the ISCAM models. You must have \code{bash}
available on your system. For Windows, you need to install WSL2
(Windows Subsystem for Linux) and install ISCAM on it to use this.

Install ISCAM: navigate to /usr/bin and clone the gfiscam GitHub
repository. Type \verb{make -j 8 dist}. It should be at:
\verb{/usr/bin/gfiscam/build/dist/bin/iscam}. This full path is necessary
because the \verb{$PATH} is not transferred through the \code{system()} call.
}
