% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{firstlower}
\alias{firstlower}
\title{Make first letter of string lowercase}
\usage{
firstlower(x)
}
\arguments{
\item{x}{string}
}
\value{
string with first letter lowercase
}
\description{
Make first letter of string lowercase
}
