% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{firstup}
\alias{firstup}
\title{Capitalize first letter of string}
\usage{
firstup(x)
}
\arguments{
\item{x}{string}
}
\value{
string with first letter capitalized
}
\description{
Capitalize first letter of string
}
