% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{fix.m}
\alias{fix.m}
\title{Change the column name 'm1' to 'm' unless both 'm1' and 'm2' exist}
\usage{
fix.m(mc)
}
\arguments{
\item{mc}{A data frame of posteriors as seen in the MCMC output csv files}
}
\value{
A data frame the same as mc but with modifications (see details)
}
\description{
Change the column name 'm1' to 'm' unless both 'm1' and 'm2' exist
}
\details{
Fo column names of data frame mc:
  If m1 and m2 both exist, no change
  If only m1 exists, change the name to m
}
