% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gear_lu_table}
\alias{gear_lu_table}
\title{Create a gear lookup table showing which gear numbers are associated with
which gear names for the various type of gears}
\usage{
gear_lu_table(model, type = c("all", "age", "index", "fleet"))
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{type}{The type of gear to include in the table.}
}
\value{

}
\description{
Create a gear lookup table showing which gear numbers are associated with
which gear names for the various type of gears
}
\details{
`type` 'all' includes all gears in the model in the order they are found in
the data file
}
