% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.age.prop}
\alias{get.age.prop}
\title{Get the value at given rank}
\usage{
get.age.prop(vec, place = 1)
}
\arguments{
\item{vec}{a vector of values}

\item{rank}{1=max, 2-second highest, etc.}
}
\value{
the value at the rank and the index where it was found
}
\description{
Get the value at given rank
}
