% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.latex.name}
\alias{get.latex.name}
\title{Get a pretty version of the parameter name}
\usage{
get.latex.name(name, addToQ = 0)
}
\arguments{
\item{name}{iscam parameter to pretty-up}

\item{addToQ}{an integer to the parameter name for the q's. This is necessary
because iscam sets the q parameter names to 1, 2, 3... regardless of the
gear number. i.e. if gear 1 is a trawl fishery and gear 2 is a survey,
iscam will call q1 the survey gear. We must add 1 to it to get q2 to
accurately portray the survey gear number}
}
\value{
an R expression which represents the pretty version of the parameter name
}
\description{
Get a pretty version of the parameter name
}
