% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get.quants}
\alias{get.quants}
\title{Calculate column quantiles for a data matrix}
\usage{
get.quants(data, probs)
}
\arguments{
\item{data}{a matrix}

\item{probs}{a vector of probabilities to be passed to [stats::quantile()]}
}
\description{
Calculate column quantiles for a data matrix
}
