% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-estimated-params.R
\name{get_estimated_params}
\alias{get_estimated_params}
\title{Fetch a data frame of the estimated MCMC parameters only}
\usage{
get_estimated_params(mc)
}
\arguments{
\item{mc}{A data frame of posteriors as seen in the MCMC output csv files}
}
\value{
A data frame of estimated parameters
}
\description{
Fetch a data frame of the estimated MCMC parameters only
}
\details{
If all values in a given column are different, it is assumed that
the parameter was estimated.
}
