% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fancy_expr}
\alias{get_fancy_expr}
\title{Get a properly-typeset version of the parameter name}
\usage{
get_fancy_expr(name, subst = FALSE)
}
\arguments{
\item{name}{iscam parameter name to make fancy}

\item{subst}{Logical. If \code{TRUE}, return all \code{substitute()}
expressions which are necessary for rendering by certain functions.
If \code{FALSE}, return all \code{expression()} expressions}
}
\value{
an R expression which represents the fancy version of the parameter name
}
\description{
Get a properly-typeset version of the parameter name
}
\details{
Includes expressions which have special characters (greek)
and super/subscripts. Also italicizes estimated parameter names.
}
