% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fancy_name}
\alias{get_fancy_name}
\title{Get a properly-typeset version of the parameter name}
\usage{
get_fancy_name(name, subst = FALSE)
}
\arguments{
\item{name}{iscam parameter name to make fancy}

\item{substr}{Logical. If `TRUE`, return all `substitute()`
expressions which are necessary for rendering by certain functions.
If `FALSE`, return all `expression()` expressions}
}
\value{
an R expression which represents the fancy version of the parameter name
}
\description{
Get a properly-typeset version of the parameter name
}
\details{
Includes expressions which have special characters (greek)
and super/subscripts. Also italicizes estimated parameter names.
}
