% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fancy_names}
\alias{get_fancy_names}
\title{Get a version of iSCAM parameter names in latex math format (i.e. $name$)}
\usage{
get_fancy_names(names)
}
\arguments{
\item{name}{A vector of iSCAM parameter names to make fancy}
}
\value{
an R expression which represents the parameter name in latex math format
}
\description{
Get a version of iSCAM parameter names in latex math format (i.e. $name$)
}
\details{
Vectorized.
Includes expressions which have special characters (greek)
and super/subscripts.
}
