% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_group_parvals}
\alias{get_group_parvals}
\title{Get a set of parameter values from the models within the model group lists}
\usage{
get_group_parvals(
  model_grp_lst,
  remove_first = FALSE,
  frac_digits = 2,
  large_digits = 0
)
}
\arguments{
\item{model_grp_lst}{Example: models$bridge_grps or models$sens_grps}

\item{remove_first}{If \code{TRUE}, remove the first model within each group
(sens groups have base model as the first one)}

\item{frac_digits}{The number of digits to report for values less than one
such as depletion}

\item{large_digits}{The number of digits to report for large numbers like
spawning biomass or B0}
}
\value{
A group list of parameter values
}
\description{
Get a set of parameter values from the models within the model group lists
}
