% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_latex_name}
\alias{get_latex_name}
\title{Get a pretty version of the parameter name. Italicizes
parameter names.}
\usage{
get_latex_name(name)
}
\arguments{
\item{name}{iscam parameter to pretty-up}
}
\value{
an R expression which represents the pretty version of the parameter name
}
\description{
Get a pretty version of the parameter name. Italicizes
parameter names.
}
