% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_num_params_est}
\alias{get_num_params_est}
\title{Extract the number of parameters estimated in a model}
\usage{
get_num_params_est(model)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}
}
\value{
The number of estimated parameters
}
\description{
Extract the number of parameters estimated in a model
}
\details{
The number is based on the iSCAM PAR file, with checks in the
control file for positiev (phz) which were means they were estimated
and not fixed
}
