% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_parvals}
\alias{get_parvals}
\title{Get a vector of values (median, CI interval, and end year) for
a parameter of your choice}
\usage{
get_parvals(model, param, digits = 0)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{param}{The name of a parameter}

\item{digits}{The number of digits to return in all numerical output}
}
\value{
A vector of 3 values, as stated in the decription
}
\description{
Get a vector of values (median, CI interval, and end year) for
a parameter of your choice
}
