% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-proj-depl.R
\name{get_proj_biomass_raw}
\alias{get_proj_biomass_raw}
\title{Extract the projection biomass or depletion columns of raw output}
\usage{
get_proj_biomass_raw(model, calc_depl = TRUE, rm_yrs = NULL, ...)
}
\arguments{
\item{model}{An ISCAM model object (class \link{mdl_cls})}

\item{calc_depl}{If \code{TRUE} calculate relative biomass (depletion) if
\code{ret_biomass_cols} is \code{TRUE}  and change the output columns to those
relative biomasses instead of absolute biomass}

\item{...}{Absorb arguments intended for other functions}
}
\value{
a data frame with the (number of posteriors * the number of catch
levels) rows and columns for each year of biomass projected with the years
as column names
}
\description{
Extract the projection biomass or depletion columns of raw output
}
