% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_rows_cols}
\alias{get_rows_cols}
\title{Calculate the number of rows and columns to have in a grid for plotting multiple figures}
\usage{
get_rows_cols(num)
}
\arguments{
\item{num}{the number of figures}
}
\value{
a vector of length 2 representing the number of rows and columns
}
\description{
Calculate the number of rows and columns to have in a grid for plotting multiple figures
}
