% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetDecade}
\alias{GetDecade}
\title{Get the decade (or other rounded value) from the year}
\usage{
GetDecade(dat, r = 10)
}
\arguments{
\item{dat}{dat}

\item{r}{r}
}
\description{
Get the decade (or other rounded value) from the year
}
