% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_iscam_model}
\alias{is_iscam_model}
\title{Determine if a model is a valid iSCAM model}
\usage{
is_iscam_model(model)
}
\arguments{
\item{model}{A proposed iscam model object}
}
\value{
Logical
}
\description{
Determine if a model is a valid iSCAM model
}
\details{
Check that the model has class `mdl_cls` and that
the attribute `model_desc` is set to something other than `NULL`
}
