% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_iscam_model_group}
\alias{is_iscam_model_group}
\title{Determine if a list is a valid iSCAM model group}
\usage{
is_iscam_model_group(lst)
}
\arguments{
\item{lst}{A proposed iscam model group object}
}
\value{
Logical
}
\description{
Determine if a list is a valid iSCAM model group
}
\details{
Checks not only that the list has class \code{mdl_grp_cls} but that each list
element is a valid iscam model list object.
}
