% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_iscam_model_list}
\alias{is_iscam_model_list}
\title{Determine if a list of models is a valid iSCAM model list}
\usage{
is_iscam_model_list(models)
}
\arguments{
\item{model}{A proposed iscam model list object}
}
\value{
Logical
}
\description{
Determine if a list of models is a valid iSCAM model list
}
\details{
Checks not only that the list has class `mdl_lst_cls` but that each list
element is a valid iscam model object.
}
