% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{iscam.exe.file}
\alias{iscam.exe.file}
\title{iscam executable filename}
\format{An object of class \code{character} of length 1.}
\usage{
iscam.exe.file
}
\description{
Executable file
}
\keyword{datasets}
