% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{iscam.starter.file}
\alias{iscam.starter.file}
\title{iscam input data filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
iscam.starter.file
}
\description{
Starter file
}
\keyword{datasets}
