% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{iscam_palette}
\alias{iscam_palette}
\title{The iSCAM model palette to use for plots by default}
\format{
An object of class \code{character} of length 1.
}
\usage{
iscam_palette
}
\description{
This variable is the default palette to use for all plots in this
package. It must be defined as one of the list given by
\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}
}
\keyword{datasets}
