% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-likelihood.R
\name{likelihood_table}
\alias{likelihood_table}
\title{Create a table of likelihood outputs}
\usage{
likelihood_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = NULL,
  ...
)
}
\arguments{
\item{models}{an iscam model object}

\item{digits}{number of decimal places for the values}

\item{french}{If \code{TRUE} translate to French}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
An \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of likelihood outputs
}
