% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{linux.pipe.to.log}
\alias{linux.pipe.to.log}
\title{The Linux (bash) pipe to log operator string}
\format{
An object of class \code{character} of length 1.
}
\usage{
linux.pipe.to.log
}
\description{
Used to pipe output to a log file when running bash scripts from  within
R
}
\keyword{datasets}
