% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load.models}
\alias{load.models}
\title{Load iscam models found in the given directory and return as a list of iscam model objects}
\usage{
load.models(path_names, inc_retro = FALSE)
}
\arguments{
\item{path_names}{a vector of the directory names for the models}

\item{inc_retro}{recursively load the retrospectives directory}
}
\value{
A list of iscam model objects
}
\description{
Load iscam models found in the given directory and return as a list of iscam model objects
}
