% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-agefit-file.R
\name{load_agefit}
\alias{load_agefit}
\title{Load the specially-formatted age fit or age residuals file}
\usage{
load_agefit(fn, type = "fits", gear_names = NULL)
}
\arguments{
\item{fn}{The filename}

\item{type}{Either 'fits' or 'resids' for the age fits (A_hat in iscam)
or residuals (A_nu in iscam)}
}
\value{
The list of output data frames representing the age fits
by gear and year
}
\description{
Load the specially-formatted age fit or age residuals file
}
