% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-longer.R
\name{load_longer}
\alias{load_longer}
\title{Load CSV files in the 'longer' data format. These files will have a
column for gear, posterior number, and year and many rows. MCMC
burnin and thinning is applied to each unique group}
\usage{
load_longer(
  fn,
  type = c("age", "sel"),
  gear_names = NULL,
  burnin = 1000,
  thin = 1,
  end_yr = year(Sys.Date()) - 1
)
}
\arguments{
\item{fn}{The CSV filename}

\item{type}{Type of file: 'age' or 'sel'. Used to set the correct start
year column name only}

\item{gear_names}{A vector of gear names to use in the resulting data frame.
If \code{NULL}, the gear numbers in the file will be used}

\item{burnin}{The number of MCMC records to remove for burnin period}

\item{thin}{Remove every nth record for thinning of MCMC output}

\item{end_yr}{The year to set as the end year for the final year block in
selectivity. If there is only one block, this will be the end year for it.
This is only used if \code{type} is "sel"}
}
\value{
The list of output data frames representing the age fits, age
residuals, or selectivity parameter estimates by gear and year
}
\description{
Load CSV files in the 'longer' data format. These files will have a
column for gear, posterior number, and year and many rows. MCMC
burnin and thinning is applied to each unique group
}
\details{
Made to load the age fit, age residuals, or selectivity
estimates files.
}
