% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rds-file.R
\name{load_rds_file}
\alias{load_rds_file}
\title{Load RDS file which was generated by \code{\link[=create_rds_file]{create_rds_file()}}}
\usage{
load_rds_file(model_dir = NULL)
}
\arguments{
\item{model_dir}{Directory name of model to be loaded}
}
\value{
The results of \code{\link[base:readRDS]{base::readRDS()}} or \code{NULL} if the file does not exist
}
\description{
Load RDS file which was generated by \code{\link[=create_rds_file]{create_rds_file()}}
}
