% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-selfit-file.R
\name{load_selest}
\alias{load_selest}
\title{Load the specially-formatted selectivity estimates file. Loading uses
parallelism by gear}
\usage{
load_selest(fn, gear_names = NULL, burnin = 1000, thin = 1)
}
\arguments{
\item{fn}{The filename}

\item{gear_names}{A vector of gear names to use in the resulting data frame.
If \code{NULL}, the gear numbers set in the age comp section of the iscam
data file will be used}

\item{burnin}{The number of MCMC records to remove for burnin period}

\item{thin}{Remove every nth record for thinning of MCMC output}
}
\value{
The list of output data frames representing the selectivity estimates
by gear and year
}
\description{
Load the specially-formatted selectivity estimates file. Loading uses
parallelism by gear
}
