% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LoadShapefiles}
\alias{LoadShapefiles}
\title{Load shapefiles: land, stat areas, etc}
\usage{
LoadShapefiles(where, a, bMax = 5000)
}
\arguments{
\item{where}{Where}

\item{a}{a}

\item{bMax}{bMax}
}
\description{
Load shapefiles: land, stat areas, etc
}
