% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{log.file}
\alias{log.file}
\title{The name of the log file used when running scripts}
\format{
An object of class \code{character} of length 1.
}
\usage{
log.file
}
\description{
The name of the log file used when running scripts
}
\keyword{datasets}
