% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age.R
\name{make.age.comp.estimates.plot}
\alias{make.age.comp.estimates.plot}
\alias{make.age.comp.fit.plot}
\alias{make.age.comp.residuals.plot}
\alias{make.age.comp.data.plot}
\title{Plot age comp estimates}
\usage{
make.age.comp.estimates.plot(model, which.gear, fg = gray(level = 0.1,
  alpha = 0.5), bg = gray(level = 0.5, alpha = 0.5), inches = 0.12)

make.age.comp.fit.plot(model, which.gear, fg = gray(level = 0.1, alpha =
  0.5), bg = gray(level = 0.5, alpha = 0.5), inches = 0.12)

make.age.comp.residuals.plot(model, which.gear, fg = gray(level = 0.1,
  alpha = 0.5), bg = gray(level = 0.5, alpha = 0.5), inches = 0.12)

make.age.comp.data.plot(model, which.gear)
}
\arguments{
\item{model}{An iscam model object}

\item{which.gear}{which gear to use. Integer index}

\item{fg}{foreground color}

\item{bg}{background color}

\item{inches}{fraction of inches to use for bubbles}
}
\description{
Plot age comp estimates

Make age comp fit plot

Make age comp residuals plot

Make age comp data plot
}
