% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{make.biomass.retro.mpd.plot}
\alias{make.biomass.retro.mpd.plot}
\title{Make a retrospective biomass plot}
\usage{
make.biomass.retro.mpd.plot(base.model, models, model.names = NULL, ylim,
  offset = 0.1, show.bo.line = FALSE, ind.letter = NULL,
  leg = TRUE, color.brew.class = "Paired", ...)
}
\arguments{
\item{base.model}{The base iscam model object}

\item{models}{A list of iscam model objects to plot against the base model}

\item{model.names}{A vector of names to show on the legend}

\item{ylim}{The y limits for the plot}

\item{offset}{The amount on the x-axis to offset each point and line for
multiple models}

\item{show.bo.line}{Show the reference lines 0.2 and 0.4bo}

\item{ind.letter}{A letter to show on the plot (for panel plots)}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{color.brew.class}{The RColorBrewer class, e.g. "Paired"}

\item{...}{Other graphical arguments}
}
\description{
Make a retrospective biomass plot
}
