% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R
\name{make.catchability.parameters.table}
\alias{make.catchability.parameters.table}
\title{Make a table of the catchability parameters - Herring specific}
\usage{
make.catchability.parameters.table(
  am1.lst,
  am2.lst = NULL,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{am1.lst}{A list of the AM1 iscam model objects}

\item{am2.lst}{A list of the AM2 iscam model objects}

\item{digits}{Number of decimal points to show}

\item{xcaption}{caption to appear in the calling document}

\item{xlabel}{the label used to reference the table in latex}

\item{font.size}{size of the font for the table}

\item{space.size}{size of the vertical spaces for the table}

\item{placement}{latex code for placement of the table in document}

\item{var}{the variance parameter results data as loaded from
variance-parameter-results.csv}

\item{which.model}{which model to make the table for, 1 = AM1, 2 = AM2}

\item{tab}{the contents of the csv file as read in by \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{model}{an iscam model object}

\item{model.am2}{an iscam model object}

\item{models}{a list of iscam model objects}

\item{model.names}{a vector of names of the models}

\item{type}{1=biomass, 2=recruitment, 3=F, 4=U, 5=depletion}

\item{syr}{start year for table}

\item{tabular.environment}{latex tabular environment value}

\item{qa.lst}{List of AM1 models for q sensitivities}

\item{qb.lst}{List of AM1 models for q sensitivities}

\item{qc.lst}{List of AM1 models for q sensitivities#'}
}
\description{
Make a table of the catchability parameters - Herring specific
}
