% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{make.catches.plot}
\alias{make.catches.plot}
\title{Plot the catches in a stacked-barplot with legend}
\usage{
make.catches.plot(catches, gear.names, leg.y.loc = 430, leg.cex = 1)
}
\arguments{
\item{catches}{output of the [load.catches()] function}

\item{gear.names}{names of the gears}

\item{leg.y.loc}{y-based location to place the legend}

\item{leg.cex}{text size for legend}
}
\description{
Plot the catches in a stacked-barplot with legend
}
