% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{make.catches.table}
\alias{make.catches.table}
\title{Make a catch xtable for latex document.}
\usage{
make.catches.table(catches, start.yr, end.yr, weight.factor = 1000,
  digits = 3, areas = NULL, xcaption = "default",
  xlabel = "default", font.size = 9, space.size = 10,
  placement = "H")
}
\arguments{
\item{catches}{output of the [load.catches()] function}

\item{start.yr}{first year to show in table}

\item{end.yr}{last year to ahow in table}

\item{weight.factor}{value to divide the catches by}

\item{digits}{number of decimal points in table}

\item{areas}{a list of vectors of areas to group together, e.g.:
areas <- list(c(3,4), c(5,6,7,8,9) means 3CD together, and 5ABCDE
together as two groups. If NULL, a summary table of catch/discard
by year will be produced}

\item{xcaption}{caption for table}

\item{xlabel}{label used as a reference in latex}

\item{font.size}{font size}

\item{space.size}{size of spaces between text in table}

\item{placement}{latex code for placement of the table in document}
}
\value{
an xtable in the proper format for catch and discards
}
\description{
Depends on the format of the catch.csv file. The function
requires that the input data frame 'catches' has at least
3 columns called 'Year', 'CatchKG', and 'DiscardedKG'
}
