% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-decisions.R
\name{make.decision.table}
\alias{make.decision.table}
\title{Produce a decision xtable for the given iscam model}
\usage{
make.decision.table(model, digits = 3, xcaption = "default",
  xlabel = "default", font.size = 10, space.size = 11,
  placement = "H")
}
\arguments{
\item{model}{An iscal model object}

\item{digits}{Number of decimal places for the values in the table}

\item{xcaption}{Caption to appear in the calling document}

\item{xlabel}{The label used to reference the table in latex}

\item{font.size}{Size of the font for the table}

\item{space.size}{Size of the vertical spaces for the table}

\item{placement}{Latex code for placement of the table in document}
}
\value{
An xtable
}
\description{
Produce a decision xtable for the given iscam model
}
