% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-indices.R
\name{make.index.fit.plot}
\alias{make.index.fit.plot}
\title{Plot the index fits for MPD runs only}
\usage{
make.index.fit.plot(models, model.names = NULL, leg = NULL, start.yr,
  end.yr, ind, ylim = c(0, 20), ind.letter = NULL, show.cv = FALSE)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{model.names}{A vector of names to show on the legend}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{start.yr}{Year to start the plot on}

\item{end.yr}{Year to end the plot on}

\item{ind}{The index to plot}

\item{ylim}{The y limits for the plot}

\item{ind.letter}{A letter to show on the plot (for panel plots)}

\item{show.cv}{Show the CV bars on the plot? TRUE/FALSE}
}
\value{
Nothing
}
\description{
Plot the index fits for MPD runs only
}
