% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-likelihood.R
\name{make.likelihood.table}
\alias{make.likelihood.table}
\title{Make an xtable of likelihood values - Herring specific}
\usage{
make.likelihood.table(
  models,
  which = 1,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 11,
  space.size = 12,
  placement = "H"
)
}
\arguments{
\item{models}{an iscam model object}

\item{which}{1 will put hg, prs, and cc in the table, 2 will put sog and wcvi in the table}

\item{digits}{number of decimal places for the values}

\item{xcaption}{caption to appear in the calling document}

\item{xlabel}{the label used to reference the table in latex}

\item{font.size}{size of the font for the table}

\item{space.size}{size of the vertical spaces for the table}

\item{placement}{latex code for placement of the table in document}
}
\description{
Make an xtable of likelihood values - Herring specific
}
