% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-maturity.R
\name{make.maturity.sens.table}
\alias{make.maturity.sens.table}
\title{Make maturity sensitivity table - Herring specific}
\usage{
make.maturity.sens.table(
  mat,
  which.model = 1,
  end.yr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{mat}{the data loaded from the maturity-sensitivity-results-table.csv file}

\item{which.model}{to make the table for, 1 = AM1, 2 = AM2}

\item{end.yr}{used for the end SB year in the table}

\item{digits}{number of decimal places for the values}

\item{xcaption}{caption to appear in the calling document}

\item{xlabel}{the label used to reference the table in latex}

\item{font.size}{size of the font for the table}

\item{space.size}{size of the vertical spaces for the table}

\item{placement}{latex code for placement of the table in document}
}
\description{
Make maturity sensitivity table - Herring specific
}
