% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-maturity.R
\name{make.maturity.table}
\alias{make.maturity.table}
\title{Make a table of maturity data}
\usage{
make.maturity.table(
  mat,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{mat}{the maturity vectors data as loaded from maturity-vectors.csv}

\item{digits}{number of decimal places for the values}

\item{xcaption}{caption to appear in the calling document}

\item{xlabel}{the label used to reference the table in latex}

\item{font.size}{size of the font for the table}

\item{space.size}{size of the vertical spaces for the table}

\item{placement}{latex code for placement of the table in document}
}
\description{
Make a table of maturity data
}
