% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-nage2.R
\name{make.nage2.plot}
\alias{make.nage2.plot}
\title{Plot the estimated number of age 2's for iscam models}
\usage{
make.nage2.plot(
  models,
  model.names = NULL,
  ylim,
  opacity = 75,
  leg = NULL,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{model.names}{A vector of names to show on the legend}

\item{ylim}{The y limits for the plot}

\item{opacity}{How opaque the credibility envelopes are}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{...}{Other graphical arguments}
}
\value{
Nothing
}
\description{
Plot the estimated number of age 2's for iscam models
}
\details{
Plot the estimated number of age 2's for iscam models.
  These are the values found in the iscam_rt_mcmc.csv file in the final year
}
