% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mortality.R
\name{make.natural.mort.mcmc.plot}
\alias{make.natural.mort.mcmc.plot}
\title{Plot the natural mortality for MCMC cases for iscam models}
\usage{
make.natural.mort.mcmc.plot(
  models,
  model.names = NULL,
  which.gear = 1,
  ylim,
  opacity = 75,
  leg = NULL,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{model.names}{A vector of names to show on the legend}

\item{which.gear}{The gear index to plot}

\item{ylim}{The y limits for the plot}

\item{opacity}{How opaque the credibility envelopes are}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{...}{Other graphical arguments}
}
\value{
Nothing
}
\description{
Plot the natural mortality for MCMC cases for iscam models
}
\details{
Plot the natural mortality with credibility intervals for the mcmc
  case of the models
}
