% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.pairs.plot}
\alias{make.pairs.plot}
\title{Plot the pairs scatterplots for the estimated parameters}
\usage{
make.pairs.plot(model, params = c("ro", "h", "m", "rbar", "rinit", "q1",
  "q2", "sel1", "selsd1", "sel2", "selsd2", "sel3", "selsd3"))
}
\arguments{
\item{model}{An iscam model object}

\item{params}{Parameters to include in plot}
}
\value{
Nothing
}
\description{
Plot the pairs scatterplots for the estimated parameters
}
