% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ref-points.R, R/table-send-params.R
\name{table_ref_points}
\alias{table_ref_points}
\alias{table_sens_params}
\title{Make a table of reference points}
\usage{
table_ref_points(
  model.am2,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)

table_sens_params(
  tab,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{translate}{Logical. Translate to french if TRUE}
}
\value{
an xtable

an xtable
}
\description{
Make a table of reference points

Make a table of the sensitivity parameter information as found in
the CSV file in the data directory
}
