% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-q.R
\name{make.q.plot}
\alias{make.q.plot}
\title{Plot the catchability trajectories for iscam models}
\usage{
make.q.plot(models, model.names = NULL, ylim, opacity = 75,
  ind.letter = NULL, leg = NULL, ...)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{model.names}{A vector of names to show on the legend}

\item{ylim}{The y limits for the plot}

\item{opacity}{How opaque the credibility envelopes are}

\item{ind.letter}{A letter to show on the plot (for panel plots)}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{...}{Other graphical arguments}
}
\value{
Nothing
}
\description{
Plot the catchability trajectories for iscam models
}
