% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-recruitment.R
\name{make.recr.mcmc.plot}
\alias{make.recr.mcmc.plot}
\title{Plot the recruitment or deviations for MCMC iscam models}
\usage{
make.recr.mcmc.plot(models, model.names = NULL, type = 1, ylim,
  offset = 0.1, append.base.txt = NULL, show.median = FALSE,
  show.mean = FALSE, ind.letter = NULL, leg = NULL, ...)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{model.names}{A vector of names to show on the legend}

\item{type}{1 = recruitment estimates, any other = recruitment deviation estimates}

\item{ylim}{The y limits for the plot}

\item{offset}{The amount on the x-axis to offset each point and line for
multiple models}

\item{append.base.txt}{Text to append to the name of the first model}

\item{show.median}{Show the median line on the plot}

\item{show.mean}{Show the mean line on the plot}

\item{ind.letter}{A letter to show on the plot (for panel plots)}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{...}{Other graphical arguments}
}
\value{
Nothing
}
\description{
Plot the recruitment or deviations for MCMC iscam models
}
\details{
Plot the recruitment or deviations with credibility intervals
  for MCMC cases for iscam models
}
