% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-recruitment.R
\name{make.recr.retro.mpd.plot}
\alias{make.recr.retro.mpd.plot}
\title{Plot the recruitment for the MPD cases for iscam models}
\usage{
make.recr.retro.mpd.plot(base.model, models, model.names = NULL, ylim,
  offset = 0.1, ind.letter = NULL, leg = TRUE,
  color.brew.class = "Paired", ...)
}
\arguments{
\item{base.model}{An iscam model object to compare against}

\item{models}{A list of iscam model objects to compare with the base model}

\item{model.names}{A vector of names to show on the legend}

\item{ylim}{The y limits for the plot}

\item{offset}{The amount on the x-axis to offset each point and line for
multiple models}

\item{ind.letter}{A letter to show on the plot (for panel plots)}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{color.brew.class}{RColorBrewer class. e.g. "Paired"}

\item{...}{Other graphical arguments}
}
\value{
Nothing
}
\description{
Plot the recruitment for the MPD cases for iscam models
}
