% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-selectivity.R
\name{make.selex.comparison.plot}
\alias{make.selex.comparison.plot}
\title{Plot the selectivity for selected gears for the iscam model}
\usage{
make.selex.comparison.plot(
  model,
  gear.names,
  show.mat = FALSE,
  bio = NULL,
  leg = NULL,
  color.brew.class = "Paired"
)
}
\arguments{
\item{model}{An iscam model object}

\item{gear.names}{A vector of the names of the gears to include.
Typically read in from the csv file in the data directory}

\item{show.mat}{plot the maturity as well}

\item{bio}{used for maturity data is show.mat is TRUE}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{color.brew.class}{RColorBrewer class. e.g. "Paired"}
}
\value{
Nothing
}
\description{
Plot the selectivity for selected gears for the iscam model
}
