% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.traces.plot}
\alias{make.traces.plot}
\title{Make trace plots for all paramaters from the mcmc output}
\usage{
make.traces.plot(model, axis.lab.freq = 200)
}
\arguments{
\item{model}{An iscam model object}

\item{axis.lab.freq}{the frequency of x-axis labelling}
}
\description{
Make trace plots for all paramaters from the mcmc output
}
