% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MakeCircle}
\alias{MakeCircle}
\title{Function to make a circle}
\usage{
MakeCircle(center = c(0, 0), radius = 1, nPts = 100)
}
\arguments{
\item{center}{center}

\item{radius}{radius}

\item{nPts}{num points}
}
\description{
Function to make a circle
}
