% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{max_na}
\alias{max_na}
\title{Calculate maximum if there are non-NA values, return NA if all values are NA}
\usage{
max_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{value}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\description{
Calculate maximum if there are non-NA values, return NA if all values are NA
}
