% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.age.fits.file}
\alias{mcmc.age.fits.file}
\title{iscam MCMC output age fits filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.age.fits.file
}
\description{
MCMC index age fits file
}
\keyword{datasets}
