% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.biomass.file}
\alias{mcmc.biomass.file}
\title{iscam MCMC output biomass filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.biomass.file
}
\description{
MCMC biomass file
}
\keyword{datasets}
