% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{mcmc.catch.file}
\alias{mcmc.catch.file}
\title{The name of the ISCAM catch file (csv)}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.catch.file
}
\description{
The name of the ISCAM catch file (csv)
}
\keyword{datasets}
