% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.file}
\alias{mcmc.file}
\title{iscam MCMC output filename}
\format{An object of class \code{character} of length 1.}
\usage{
mcmc.file
}
\description{
MCMC file
}
\keyword{datasets}
