% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.fishing.mort.file}
\alias{mcmc.fishing.mort.file}
\title{iscam MCMC output instantaneous fishing mortality filename}
\format{An object of class \code{character} of length 1.}
\usage{
mcmc.fishing.mort.file
}
\description{
MCMC F file
}
\keyword{datasets}
