% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.fishing.mort.u.file}
\alias{mcmc.fishing.mort.u.file}
\title{iscam MCMC output annual fishing mortality filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.fishing.mort.u.file
}
\description{
MCMC U file
}
\keyword{datasets}
