% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.index.fits.file}
\alias{mcmc.index.fits.file}
\title{iscam MCMC output index fits filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.index.fits.file
}
\description{
MCMC index fits file
}
\keyword{datasets}
