% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.natural.mort.file}
\alias{mcmc.natural.mort.file}
\title{iscam MCMC output natural mortality filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.natural.mort.file
}
\description{
MCMC M file
}
\keyword{datasets}
