% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.recr.devs.file}
\alias{mcmc.recr.devs.file}
\title{iscam MCMC output recruitment deviations filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.recr.devs.file
}
\description{
MCMC recruitment deviations file
}
\keyword{datasets}
