% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.recr.file}
\alias{mcmc.recr.file}
\title{iscam MCMC output recruitment filename}
\format{An object of class \code{character} of length 1.}
\usage{
mcmc.recr.file
}
\description{
MCMC recruitment file
}
\keyword{datasets}
