% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.sel.file}
\alias{mcmc.sel.file}
\title{iscam MCMC output selectivity fits filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.sel.file
}
\description{
MCMC index selectivity fits file
}
\keyword{datasets}
