% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{mcmc.sr.file}
\alias{mcmc.sr.file}
\title{The name of the ISCAM stock recruitment file (csv)}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.sr.file
}
\description{
The name of the ISCAM stock recruitment file (csv)
}
\keyword{datasets}
