% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{mcmc.thin}
\alias{mcmc.thin}
\title{Apply burnin and thinning to the MCMC posteriors}
\usage{
mcmc.thin(mcmc.dat, burnin, thin)
}
\arguments{
\item{mcmc.dat}{A data frame of the MCMC posteriors}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{thin}{The thinning to apply to the MCMC posterior samples}
}
\value{
an mcmc.window object (CODA package)
}
\description{
Apply burnin and thinning to the MCMC posteriors
}
