% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.vuln.biomass.file}
\alias{mcmc.vuln.biomass.file}
\title{iscam MCMC output vulnerable biomass filename}
\format{An object of class \code{character} of length 1.}
\usage{
mcmc.vuln.biomass.file
}
\description{
MCMC vulnerable biomass file
}
\keyword{datasets}
