% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{mdl_grp_cls}
\alias{mdl_grp_cls}
\title{The iSCAM model group class}
\format{
An object of class \code{character} of length 1.
}
\usage{
mdl_grp_cls
}
\description{
This class defines a list as being a group of one or more iSCAM model lists.
Each element of the list must have class \link{mdl_lst_cls} to qualify.
Model lists are assigned this class in \code{\link[=model_setup]{model_setup()}}
}
\keyword{datasets}
