% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mean_na}
\alias{mean_na}
\title{Calculate mean if there are non-NA values, return NA if all values are NA}
\usage{
mean_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{value}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\description{
Calculate mean if there are non-NA values, return NA if all values are NA
}
