% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{min_na}
\alias{min_na}
\title{Calculate minimum if there are non-NA values, return NA if all values are NA}
\usage{
min_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{value}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\description{
Calculate minimum if there are non-NA values, return NA if all values are NA
}
