% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{model.class}
\alias{model.class}
\title{iscam model}
\format{An object of class \code{character} of length 1.}
\usage{
model.class
}
\description{
Class definition
}
\keyword{datasets}
