% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{model.lst.class}
\alias{model.lst.class}
\title{The name of the ISCAM model class}
\format{
An object of class \code{character} of length 1.
}
\usage{
model.lst.class
}
\description{
The name of the ISCAM model class
}
\keyword{datasets}
