% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{model_setup}
\alias{model_setup}
\title{Load models and set up lists and classes for the base model, bridge model groups, and
sensitivity model groups}
\usage{
model_setup(
  main_dirs = NULL,
  bridge_models_text = NULL,
  sens_models_text = NULL,
  retro_models_text = NULL,
  ...
)
}
\arguments{
\item{main_dirs}{Output list from \code{\link[=set_dirs]{set_dirs()}}}

\item{bridge_models_text}{A list of vectors of text strings to show in the legends for bridge
model plots, one name for each model, where the list elements represent a group of models}

\item{sens_models_text}{A list of vectors of text strings to show in the legends for sensitivity
model plots, one name for each model, where the list elements represent a group of models}

\item{retro_models_text}{A list of vectors of text strings to show in the legends for retrospective
model plots, one name for each model, where the list elements represent a group of models}

\item{...}{Arguments to pass to \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A list of three items, the base_model inside a single-element list, the list of
bridge model groups, and the list of sensitivity model groups. \code{bridge_grp} and \code{sens_grp}
are groups lists of models which are to be compared with each other in the document.
This simplifies plotting and table functions
}
\description{
Load models and set up lists and classes for the base model, bridge model groups, and
sensitivity model groups
}
\examples{
\dontrun{
library(gfiscamutils)
bridge_models_dirs <- list(c("01-base-2015",
                             "02-bridge-update-data"),
                           c("03-bridge-update-likelihood",
                             "04-bridge-update-survey"))
bridge_models_text <- list(c("base model from 2015",
                             "Add new data"),
                           c("Use new lieklihood method",
                             "Add new survey index point"))
sens_models_dirs <- list(c("10-high-m",
                           "11-low-m"),
                         c("12-high-sigma-r"))
sens_models_text <- list(c("Increase prior for M",
                           "Decrease prior for M"),
                         c("Increase prior for sigma R"))
drs <- set_dirs(base_model_dir = "base",
                bridge_models_dirs = bridge_models_dirs,
                sens_model_dirs = sens_model_dirs)
retro_models_dirs <- list(c("01-retrospective-1-year",
                            "02-retrospective-2-year",
                            "03-retrospective-3-year"))
retro_models_text <- list(c("Base model - 1 year",
                            "Base model - 2 years",
                            "Base model - 3 years"))
model_setup <- function(drs,
                        bridge_models_text = bridge_models_text,
                        sens_models_text = sens_models_text,
                        retro_models_text = retro_models_text,
                        overwrite_rds_files = TRUE)
}
}
