% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modify_model_path}
\alias{modify_model_path}
\title{Read in model RDS files, modify the path attribute, and save}
\usage{
modify_model_path(model_fns, paths)
}
\arguments{
\item{model_fns}{A vector of model RDS filenames to modify}

\item{paths}{A vector of paths to replace in the files. Must be the
same length as \code{model_fns}}
}
\description{
Read in model RDS files, modify the path attribute, and save
}
