% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mpd.proj.file}
\alias{mpd.proj.file}
\title{iscam MPD projection filename}
\format{An object of class \code{character} of length 1.}
\usage{
mpd.proj.file
}
\description{
MPD projection file
}
\keyword{datasets}
