% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{output.files}
\alias{output.files}
\title{A vector of ADMB output files for ISCAM}
\format{
An object of class \code{character} of length 24.
}
\usage{
output.files
}
\description{
A vector of ADMB output files for ISCAM
}
\keyword{datasets}
