% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{par.file}
\alias{par.file}
\title{iscam parameter estimate filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
par.file
}
\description{
PAR file
}
\keyword{datasets}
