% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-selectivity.R
\name{sel_param_est_mpd_table}
\alias{sel_param_est_mpd_table}
\title{Make a table of selectivity parameter estimate comparisons for MPD models}
\usage{
sel_param_est_mpd_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = NULL,
  ...
)
}
\value{
an [csasdown::csas_table()]
}
\description{
Make a table of selectivity parameter estimate comparisons for MPD models
}
