% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R
\name{param_settings_table}
\alias{param_settings_table}
\title{Make a table showing number of parameters estimated and prior parameterizations}
\usage{
param_settings_table(model, french = FALSE, ...)
}
\arguments{
\item{model}{A single output model}

\item{french}{If \code{TRUE} translate to French}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\description{
Make a table showing number of parameters estimated and prior parameterizations
}
