% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PasteNicely}
\alias{PasteNicely}
\title{Paste strings nicely}
\usage{
PasteNicely(x, intChars = ", ", nChar = "and ")
}
\arguments{
\item{x}{vector of strings}

\item{intChars}{Seperator}

\item{nChar}{Last seperator}
}
\description{
Paste strings nicely
}
