% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{plot.marg}
\alias{plot.marg}
\title{Plot posterior histograms}
\usage{
\method{plot}{marg}(xx, breaks = "sturges", ex.factor = 1, ...)
}
\arguments{
\item{xx}{A list containing values used to plot a posterior histogram.
See [make.biomass.mcmc.plot()] to see how this is constructed}

\item{breaks}{Same as the 'breaks' argument in [graphics::hist()]}

\item{ex.factor}{A factor to change the x-axis range}

\item{...}{Other graphical arguments}
}
\description{
Plot posterior histograms
}
