% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-fits-mcmc.R
\name{plot_age_fits_mcmc}
\alias{plot_age_fits_mcmc}
\title{Plot MCMC age fits for an iSCAM model}
\usage{
plot_age_fits_mcmc(
  model,
  gear = 1,
  yrs = NULL,
  comp_color = "black",
  comp_width = 0.5,
  comp_alpha = 0.5,
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  text_title_size = 12,
  text_title_inc_mdl_nm = FALSE,
  angle_x_labels = FALSE,
  show_sample_size_f_only = TRUE,
  sample_size_x = 17,
  sample_size_y = 0.2
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{yrs}{A vector of years to include in the plot. If the maximum extends
past the range of years in the data, the maximum in the data will be used.
If \code{NULL}, all years will be included}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_width}{Age comp bar widths}

\item{comp_alpha}{Age comp bar alpha value (0-1)}

\item{ci_linetype}{See \code{linetype} in \link{ggplot2}. Only used if \code{ci_type} is
"line" or "both"}

\item{ci_color}{Color for the lines or ribbon for the credible intervals}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons.
Only used if \code{ci_type} is "ribbon" or "both"}

\item{text_title_size}{Size of text for the axis titles}

\item{text_title_inc_mdl_nm}{Logical. If \code{TRUE}, make the model name the
main title and the gear the subtitle. If \code{FALSE}, do not show the model
name, and make the gear the main title (if \code{text_title_size} is \code{NULL},
no title will be shown)}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}

\item{show_sample_size_f_only}{If \code{TRUE}, show the sample size in the female
panels only. If \code{FALSE}, show the sample size in both male and female panels.
Included for cases where the sample size is shared or repeated for both sexes.}

\item{sample_size_x}{The x value (age) to place the sample size label}

\item{sample_size_y}{The y value (0-1) to place the sample size label}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot MCMC age fits for an iSCAM model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mpd}()},
\code{\link{plot_age_resids_boxes_mcmc}()},
\code{\link{plot_age_resids_mcmc}()},
\code{\link{plot_age_resids_mpd}()}
}
\concept{Age plotting functions}
