% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-fits-mpd.R
\name{plot_age_fits_mpd}
\alias{plot_age_fits_mpd}
\title{Plot MPD age fits for an iSCAM model}
\usage{
plot_age_fits_mpd(
  model,
  gear = 1,
  comp_color = "black",
  comp_point_size = 0.5,
  fit_line_color = "red",
  text_title_size = 12,
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age comp data points}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot MPD age fits for an iSCAM model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mcmc}()},
\code{\link{plot_age_resids_boxes_mcmc}()},
\code{\link{plot_age_resids_mcmc}()},
\code{\link{plot_age_resids_mpd}()}
}
\concept{Age plotting functions}
