% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-fits-mpd.R,
%   R/figures-age-residuals-mpd.R
\name{plot_age_fits_mpd}
\alias{plot_age_fits_mpd}
\alias{plot_age_resids_mpd}
\title{Plot the age fits for an MPD model}
\usage{
plot_age_fits_mpd(
  model,
  gear = 1,
  comp_color = "black",
  comp_point_size = 0.5,
  fit_line_color = "red",
  text_title_size = 12,
  angle_x_labels = FALSE
)

plot_age_resids_mpd(
  model,
  gear = 1,
  type = c("age", "year", "birth_year"),
  comp_color = "black",
  comp_point_size = 0.5,
  fir_line_color = "red",
  ylim = c(-3, 3),
  text_title_size = 12,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\value{
A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object
}
\description{
Plot the age fits for an MPD model

Plot the age fits for an MPD model
}
