% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-resids-boxes-mcmc.R
\name{plot_age_resids_boxes_mcmc}
\alias{plot_age_resids_boxes_mcmc}
\title{Plot MCMC age residuals for an iSCAM model}
\usage{
plot_age_resids_boxes_mcmc(
  model,
  gear = 1,
  type = c("age", "year", "birth_year"),
  probs = c(0.025, 0.5, 0.975),
  comp_color = "black",
  comp_point_size = 0.5,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  ylim = c(-3, 3),
  text_title_size = 12,
  text_title_inc_mdl_nm = FALSE,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{type}{One of "age", "year", or "birth_year"}

\item{probs}{Probability values to use in the plots}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age composition points}

\item{ci_type}{One of "both" or "line", type of confidence interval}

\item{ci_linetype}{See \code{linetype} in \link{ggplot2}. Only used if \code{ci_type} is
"line" or "both"}

\item{ci_color}{Color for the lines or ribbon for the credible intervals}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons.
Only used if \code{ci_type} is "ribbon" or "both"}

\item{ylim}{The y-axis limits (vector of 2)}

\item{text_title_size}{Size of text for the axis titles}

\item{text_title_inc_mdl_nm}{Logical. If \code{TRUE}, make the model name the
main title and the gear the subtitle. If \code{FALSE}, do not show the model
name, and make the gear the main title (if \code{text_title_size} is \code{NULL},
no title will be shown)}

\item{leg_loc}{Legend location (vector of 2)}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot MCMC age residuals for an iSCAM model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mcmc}()},
\code{\link{plot_age_fits_mpd}()},
\code{\link{plot_age_resids_mcmc}()},
\code{\link{plot_age_resids_mpd}()}
}
\concept{Age plotting functions}
