% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-resids-mcmc.R
\name{plot_age_resids_mcmc}
\alias{plot_age_resids_mcmc}
\title{Plot MCMC age residuals bubbles for an iSCAM model}
\usage{
plot_age_resids_mcmc(
  model,
  gear = 1,
  include = c(-5, 5),
  text_title_size = 12,
  text_title_inc_mdl_nm = FALSE,
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{include}{Two-element vector of the minimum and maximum values
to include. Eliminates large outliers which bias the bubble sizes}

\item{text_title_size}{Size of text for the axis titles}

\item{text_title_inc_mdl_nm}{Logical. If \code{TRUE}, make the model name the
main title and the gear the subtitle. If \code{FALSE}, do not show the model
name, and make the gear the main title (if \code{text_title_size} is \code{NULL},
no title will be shown)}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot MCMC age residuals bubbles for an iSCAM model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mcmc}()},
\code{\link{plot_age_fits_mpd}()},
\code{\link{plot_age_resids_boxes_mcmc}()},
\code{\link{plot_age_resids_mpd}()}
}
\concept{Age plotting functions}
