% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-resids-mpd.R
\name{plot_age_resids_mpd}
\alias{plot_age_resids_mpd}
\title{Plot the age fits for an MPD model}
\usage{
plot_age_resids_mpd(
  model,
  gear = 1,
  type = c("age", "year", "birth_year"),
  comp_color = "black",
  comp_point_size = 0.5,
  fit_line_color = "red",
  ylim = c(-3, 3),
  text_title_size = 12,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{type}{One of "age", "year", or "birthy_year"}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age composition points}

\item{fit_line_color}{Color for the age fit line}

\item{ylim}{The y-axis limits (vector of 2)}

\item{text_title_size}{Size of text for the axis titles}

\item{leg_loc}{Legend location (vector of 2)}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the age fits for an MPD model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mcmc}()},
\code{\link{plot_age_fits_mpd}()},
\code{\link{plot_age_resids_boxes_mcmc}()},
\code{\link{plot_age_resids_mcmc}()}
}
\concept{Age plotting functions}
