% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age.R
\name{plot_age_year_birthyear_residuals}
\alias{plot_age_year_birthyear_residuals}
\title{Plot a three-panel residual figure. The panels are Age, Year, and Birth year residuals}
\usage{
plot_age_year_birthyear_residuals(
  model,
  model_name = "iSCAM model",
  gear = 1,
  gear_name = model$dat$age_gear_abbrevs[gear],
  ylim_age = c(-1, 1),
  ylim_year = c(-1, 1),
  ylim_birthyear = c(-1, -1),
  legend_pos = c(0.05, 0.15),
  show_title = TRUE,
  sngle_x_labels = C(FALSE, FALSE, TRUE)
)
}
\arguments{
\item{model}{An iSCAM model}

\item{model_name}{Name of the model for plot title}

\item{gear}{Number of gear to plot}

\item{gear_name}{Name of the gear for plot label}

\item{ylim_age}{ylim for the Age plot}

\item{ylim_year}{ylim for the Year plot}

\item{ylim_birthyear}{ylim for the birthyear plot}

\item{legend_pos}{Position of the legend inside the plot area (two-element vector of x, y)}

\item{show_title}{If `TRUE`, show the title (model and gear names) on top of the three panels}

\item{angle_x_labels}{A three-element vector of logicals. If `TRUE`, angle the x-axis tick
labels for that plot, where index 1 = Age, 2 = Year, and 3 = Birthyear plots}
}
\value{
a [cowplot::plot_grid()] object
}
\description{
Plot a three-panel residual figure. The panels are Age, Year, and Birth year residuals
}
