% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age.R
\name{plot_agecomp_fits_mpd}
\alias{plot_agecomp_fits_mpd}
\title{Plot MPD age comp data along with the model fits for models
with both sexes}
\usage{
plot_agecomp_fits_mpd(
  model,
  gear,
  model_name = "",
  type = "bars",
  pearson_type = "age",
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{Model list as output by [model_setup()]}

\item{gear}{The gear number to plot}

\item{type}{One of 'bars' or 'pearson'}

\item{pearson_type}{One of 'age', 'year', or 'birthyear'}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Plot MPD age comp data along with the model fits for models
with both sexes
}
