% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age.R
\name{plot_agecomp_fits_splitsex}
\alias{plot_agecomp_fits_splitsex}
\title{Plot MPD age comp data along with the model fits for models
with both sexes}
\usage{
plot_agecomp_fits_splitsex(model, gear)
}
\arguments{
\item{model}{Model list as output by [model_setup()]}

\item{gear}{The gear number to plot}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Plot MPD age comp data along with the model fits for models
with both sexes
}
