% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-mcmc.R
\name{plot_agefits_mcmc}
\alias{plot_agefits_mcmc}
\title{Plot the age fits for an MCMC model}
\usage{
plot_agefits_mcmc(
  model,
  gear = 1,
  type = c("fits", "resids"),
  resid_type = c("age", "year", "birth_year"),
  probs = c(0.025, 0.5, 0.975),
  comp_color = "black",
  comp_point_size = 0.5,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{gear}{The number of the gear to plot}

\item{type}{The type of plot to make. Options are "fits" and "resids"}

\item{resid_type}{The type of residual plot to make if `type == "resids"`.
Options are "age", "year", and "birth_year"}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age comp data points}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See `linetype` in [ggplot2]. Only used if `ci_type` is "line" or "both"}

\item{ci_color}{Color for the lines or ribbon for the credible intervals}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons. Only used if
`ci_type` is "ribbon" or "both"}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the age fits for an MCMC model
}
