% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-mpd.R
\name{plot_agefits_mpd}
\alias{plot_agefits_mpd}
\title{Plot the age fits for an MPD model}
\usage{
plot_agefits_mpd(
  model,
  gear = 1,
  type = c("fits", "resids"),
  resid_type = c("age", "year", "birth_year"),
  comp_color = "black",
  comp_point_size = 0.5,
  fir_line_color = "red",
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{gear}{The number of the gear to plot}

\item{resid_type}{The type of residual plot to make if `type == "resids"`.
Options are "age", "year", and "birth_year"}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age comp data points}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}

\item{fit_line_color}{Color for the fit line}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the age fits for an MPD model
}
