% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-autocor.R
\name{plot_autocor}
\alias{plot_autocor}
\title{Plot autocorrelation plot for all parameters in an MCMC run}
\usage{
plot_autocor(model, plot_sel = FALSE, param_rm = NULL, lag = 100, ...)
}
\arguments{
\item{model}{An iscam model}

\item{plot_sel}{Logical. If `TRUE`, plot only the selectivity parameters.
If `FALSE`, plot only non-selectivity parameters}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If `NULL`, nothing is removed}

\item{...}{Arguments passed to [cowplot::plot_grid()]}
}
\value{
A [ggplot2::ggplot()] object made from the [forecast::ggAcf()] function
}
\description{
Plot autocorrelation plot for all parameters in an MCMC run
}
