% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass-grid-mcmc.R
\name{plot_biomass_grid_mcmc}
\alias{plot_biomass_grid_mcmc}
\title{Plot a grid of MCMC biomass plots for iSCAM models}
\usage{
plot_biomass_grid_mcmc(models, ...)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} object
}
\description{
Plot a grid of biomass plots, each with one model only so that the reference
point lines and credible intervals can be compared correctly.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_biomass_proj_mcmc}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
