% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass-grid-mcmc.R,
%   R/figures-biomass-mcmc.R, R/figures-biomass-mpd.R, R/figures-recr-mpd.R,
%   R/figures-recruitment-mcmc.R
\name{plot_biomass_grid_mcmc}
\alias{plot_biomass_grid_mcmc}
\alias{plot_biomass_mcmc}
\alias{plot_biomass_mpd}
\alias{plot_recr_mpd}
\alias{plot_recr_mcmc}
\title{Plot a grid of biomass plots, each with one model only so that
the B0 and/or the BMSY lines and credible intervals can be viewed}
\usage{
plot_biomass_grid_mcmc(models, ...)

plot_biomass_mcmc(
  models,
  rel = FALSE,
  show_bo = TRUE,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  first_model_ribbon = TRUE,
  refpts_ribbon = TRUE,
  alpha = 0.2,
  refpts_alpha = alpha,
  palette = "Paired",
  base_color = "black",
  bo_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  angle_x_labels = FALSE,
  ...
)

plot_biomass_mpd(
  models,
  rel = FALSE,
  show_bo = TRUE,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  alpha = 0.2,
  palette = "Paired",
  base_color = "black",
  bo_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ind_letter = NULL,
  leg_loc = NULL,
  angle_x_labels = FALSE,
  ...
)

plot_recr_mpd(
  models,
  model_names = NULL,
  show_ro = TRUE,
  ro_color = base_color,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  alpha = 0.2,
  palette = "Paired",
  base_color = "#000000",
  r_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  angle_x_labels = FALSE,
  ...
)

plot_recr_mcmc(
  models,
  model_names = NULL,
  show_ro = TRUE,
  ro_color = base_color,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  ro_ribbon = TRUE,
  ro_alpha = 0.3,
  palette = "Paired",
  base_color = "#000000",
  r_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  angle_x_labels = FALSE,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class [mdl_lst_cls])}

\item{...}{Other graphical arguments}

\item{rel}{Logical. Make plot relative to initial estimate (B0), also known as depletion}

\item{show_bo}{Logical. If `TRUE` and `rel == FALSE`, show the initial value
on the plot (B0)}

\item{legend_title}{Title for legend}

\item{xlim}{The x limits for the plot. If `NULL`, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If `NULL`, the limits of the data
will be used}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{first_model_ribbon}{Logical. If `TRUE`, give the first model a shaded
credible interval instead of dotted lines}

\item{refpts_ribbon}{Logical. If `TRUE`, make the first model's reference points lines
(`show_bo_lines` and/or `show_bmsy_lines` must be `TRUE`) plotted an envelope
of the credible interval, surrounding the median lines for the reference points}

\item{alpha}{The opacity between 0 to 1 of the envelope shown when `first_model_ribbon == TRUE`}

\item{refpts_alpha}{The opacity between 0 to 1 of the envelope shown for referece points
when `refpts_ribbon == TRUE` and `show_bo_lines` and/or `show_bmsy_lines` are `TRUE`}

\item{palette}{A palette value that is in [RColorBrewer::brewer.pal.info]}

\item{base_color}{A color to prepend to the brewer colors which are set by `palette`.
This is called `base_color` because it is likely to be a base model}

\item{bo_dodge}{The amount to offset the initial value (B0 or R0) values from each
other so the values and uncertainty can be easily seen for multiple models}

\item{x_space}{The amount of x-interval space to pad the left and right of the plot
with. To remove all padding, make this 0}

\item{append_base_txt}{Text to append to the first model's name for display on the
plot legend}

\item{show_bo_lines}{Show the B0 lines at values given by `bo_refpts` for the
first model in the `models` list}

\item{bo_refpts}{Vector of two proportional values for the limit reference point
and Upper stock reference. Values are 0.2B0 and 0.4B0 by default}

\item{show_bmsy_lines}{Show the BMSY lines at values given by `bmsy_refpts` for the
first model in the `models` list}

\item{bmsy_refpts}{Vector of two proportional values for the limit reference point
and Upper stock reference. Values are 0.4BMSY and 0.8BMSY by default}

\item{bo_refpt_colors}{A vector of two colors representing the LRP and USR for B0.
Used to display reference point lines if `show_bo_lines == TRUE`}

\item{bmsy_refpt_colors}{A vector of two colors representing the LRP and USR for BMSY.
Used to display reference point lines if `show_bmsy_lines == TRUE`}

\item{ind_letter}{A letter to place in the upper left corner of the plot. If `NULL`,
nothing will be shown}

\item{leg_loc}{See the `legend.position` argument in [ggplot2::theme()]. Use "none"
to remove legend completely}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}

\item{show_ro}{Show the initial recruitment, R0 median line and credible interval}

\item{r_dodge}{See `bo_dodge` in [plot_biomass_mcmc()]}

\item{ro_ribbon}{See `refpts_ribbon` in [plot_biomass_mcmc()]}

\item{ro_alpha}{See `refpts_alpha` in [plot_biomass_mcmc()]}
}
\value{
A [cowplot::plot_grid()] object

A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object
}
\description{
Plot a grid of biomass plots, each with one model only so that
the B0 and/or the BMSY lines and credible intervals can be viewed

Plot the MCMC spawning biomass trajectories for iscam models in either
absolute or relative form.

Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.

Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.

Plot the MCMC recruitment time series trajectories for iscam models.
}
\details{
The model description text will appear in the legend if `models` length
is greater than 1, and as a title if there is only one model. The
description text is found in the `model_desc` attribute of a
`mdl_cls` class ("iscam_model") object. The attribute is set in the
[model_setup()] function, from either the `bridge_model_text` or
`sens_model_text` arguments.

This function cannot have more than one model with a shaded credible interval.
It is too difficult to tell what's going on with colors overlapping.
If called with `rel = TRUE`, `show_bo` will be overridden to be `FALSE`.
It makes no sense to show B0 when a relative plot is asked for, it is always 1.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}
}
\concept{Biomass plotting functions}
\concept{Time series plotting functions}
