% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass-mpd.R
\name{plot_biomass_mpd}
\alias{plot_biomass_mpd}
\title{Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.}
\usage{
plot_biomass_mpd(
  models,
  model_names = NULL,
  type = c("sbt", "rt"),
  rel = FALSE,
  show_initial = TRUE,
  legend_title = "Models",
  palette = "Paired",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  line_ribbon = FALSE,
  refpts_ribbon = TRUE,
  alpha = 0.2,
  refpts_alpha = alpha,
  offset = 0.1,
  bo_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bo_lines = FALSE,
  show_bmsy_lines = FALSE,
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  ...
)
}
\value{
Nothing
}
\description{
Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.
}
