% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass-proj-mcmc.R
\name{plot_biomass_proj_mcmc}
\alias{plot_biomass_proj_mcmc}
\title{Plot a model's time series (either spawning biomass or relaive spawning
biomass) and its projections}
\usage{
plot_biomass_proj_mcmc(
  model,
  model_name = "Base model",
  rel = FALSE,
  probs = c(0.025, 0.5, 0.975),
  x_space = 0.5,
  y_space = ifelse(rel, 0.05, 0.5),
  show_bo_lines = FALSE,
  bo_refpt_colors = c("red", "green"),
  line_width = 1,
  point_size = 2,
  label_digits = 2,
  label_append_text = " t",
  ...
)
}
\arguments{
\item{model}{An ISCAM model object (class \link{mdl_cls})}

\item{model_name}{Name of the input model}

\item{rel}{Logical. Make plot relative to initial estimate (B0), also known
as depletion}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{y_space}{The amount of y-interval space to pad the top and bottom of
the plot with. To remove all padding, make this 0}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{label_digits}{Number of digits to show in the catch labels}

\item{label_append_text}{Text to append to the labels}

\item{...}{Arguments passed to \code{\link[=plot_ts_mcmc]{plot_ts_mcmc()}}}
}
\description{
Plot a model's time series (either spawning biomass or relaive spawning
biomass) and its projections
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
