% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch-fleet.R
\name{plot_catch_fleet}
\alias{plot_catch_fleet}
\title{Plot catch by fleet}
\usage{
plot_catch_fleet(catch_lst, xlim)
}
\arguments{
\item{catch_lst}{A list of catch data frames, as output from
\code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{xlim}{The x-axis limits (vector of two)}
}
\value{
Nothing
}
\description{
Plot catch by fleet
}
