% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dfo-ref-point-figure.R
\name{plot_dfo_ref_point_figure}
\alias{plot_dfo_ref_point_figure}
\title{Plot DFO-style reference point figure with horizontal boxplots and zones
bounded by B0 reference points}
\usage{
plot_dfo_ref_point_figure(
  model,
  biomass_col = "B2023",
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975)
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{biomass_col}{The colomn name that you want to shoe output for.
This column will be divided by the SB0 value to get a depletion for
the year in the name of the column}
}
\description{
Plot DFO-style reference point figure with horizontal boxplots and zones
bounded by B0 reference points
}
