% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-f-mcmc.R
\name{plot_f_mcmc}
\alias{plot_f_mcmc}
\title{Plot MCMC fishing mortalities for iSCAM models}
\usage{
plot_f_mcmc(..., type = c("i", "e"))
}
\arguments{
\item{...}{Absorbs arguments meant for other functions}

\item{type}{'i' means instantaneous, 'e' means exploitation rate (Ut)}
}
\description{
Plot fishing mortality time series plots for MCMC output with credible
intervals
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_biomass_proj_mcmc}()},
\code{\link{plot_catch_fit_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
