% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-index-mpd.R
\name{plot_index_mpd}
\alias{plot_index_mpd}
\title{Plot MPD index fits or residuals for iSCAM models}
\usage{
plot_index_mpd(
  models,
  type = c("fits", "resids"),
  surv_index,
  gear = NULL,
  start_year = 1995,
  end_year = 2021,
  append_base_txt = NULL,
  legend_title = "Models",
  palette = "Paired",
  base_color = "black",
  dodge = 0.3,
  index_line_width = 0.5,
  index_point_size = 2,
  index_color = "chocolate3",
  fit_line_width = 0.5,
  fit_point_size = 2,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{type}{Either 'fits' or 'resids' for model fits or residuals respectively}

\item{surv_index}{The \code{survey_index} data frame which is \code{dat$survey_index}
if \code{dat} is the output from the \code{\link[gfdata:get_data]{gfdata::get_survey_index()}} function}

\item{gear}{A vector of gear numbers to show panels for. If \code{NULL}, all will be shown.
If a gear number out of range is included, a lookup table with gear numbers and names
will be shown}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{legend_title}{Title for legend}

\item{palette}{A palette value that is in \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}}

\item{base_color}{A color to prepend to the brewer colors which are set by
\code{palette}. This is called \code{base_color} because it is likely to be a base
model}

\item{index_line_width}{The index data error bar and connecting line width}

\item{index_point_size}{The index data point size}

\item{index_color}{The color used for the observed index lines and points}

\item{fit_line_width}{The model fit error bar and connecting line width}

\item{fit_point_size}{The model fit point size}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner. It can also be the string "facet" in which case
the legend will appear in the empty facet if it exists.}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\description{
Plot MPD index fits or residuals for a single or group of models by gear
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
