% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-indices.R
\name{plot_index_resids_mpd}
\alias{plot_index_resids_mpd}
\title{Plot the residuals for multiple models}
\usage{
plot_index_resids_mpd(models, gear = "SYN QCS")
}
\arguments{
\item{models}{A list of iSCAM models}

\item{gear}{One of "SYN QCS", "SYN HS", "SYN WCVI", "SYN WCHG", "DCPUE"}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the residuals for multiple models
}
