% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-normal-curve.R
\name{plot_norm}
\alias{plot_norm}
\title{Plot a normal curve}
\usage{
plot_norm(
  nsamp = 100,
  mean = 0,
  sd = 0.5,
  xlim = c(-6, 6),
  fill = "steelblue",
  mean_line_col = "white",
  mean_line_lwd = 0.5,
  mean_line_lty = 1,
  sd_lines_col = "white",
  sd_lines_lwd = 0.5,
  sd_lines_lty = 2,
  alpha = 0.5
)
}
\arguments{
\item{nsamp}{Number of points across x-axis}

\item{mean}{Mean of normal}

\item{sd}{Standard deviation of normal}

\item{xlim}{x-axis limits (2-element vector)}

\item{fill}{Fill color for under the curve}

\item{mean_line_col}{Mean line color}

\item{mean_line_lwd}{Mean line width}

\item{mean_line_lty}{Mean line type}

\item{sd_lines_col}{Standard deviation lines color}

\item{sd_lines_lwd}{Standard deviation lines width}

\item{sd_lines_lty}{Standard deviation lines type}

\item{alpha}{Opacity of fill from 0-1}
}
\value{
a [ggplot2::ggplot()] object
}
\description{
Plot a normal curve
}
