% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age.R
\name{plot_pa}
\alias{plot_pa}
\title{Plot proportions-at-age time series from a data frames as extracted from iscam data (dat) files}
\usage{
plot_pa(df, age_plus = 10, conf = 0.9, xlim = c(1000, 3000),
  ylim = c(0, NA), translate = FALSE)
}
\arguments{
\item{df}{a data frame as constructed by [get_pa()]}

\item{age_plus}{age plus group}

\item{conf}{confidence value for the envelope}

\item{xlim}{limits for the years shown on the plot}

\item{ylim}{limits for the ages shown on the plot}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
A ggplot object
}
\description{
Plot proportions-at-age time series from a data frames as extracted from iscam data (dat) files
}
