% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Plot the estimated parameters in a model against each other
as a pairs plot}
\usage{
plot_pairs(model, plot_sel = FALSE, param_rm = NULL)
}
\arguments{
\item{model}{An iscam model}
}
\value{
A [ggplot2::ggplot()] object made from the [GGally::ggpairs()] function
}
\description{
Plot the estimated parameters in a model against each other
as a pairs plot
}
