% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pairs-mcmc.R
\name{plot_pairs_mcmc}
\alias{plot_pairs_mcmc}
\title{Plot MCMC pairs for iSCAM models}
\usage{
plot_pairs_mcmc(
  model,
  plot_sel = NULL,
  param_rm = c("rho", "vartheta"),
  list_param_names = FALSE,
  text_title_size = 12
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{plot_sel}{Logical. If \code{TRUE}, plot only the selectivity parameters.
If \code{FALSE}, plot only non-selectivity parameters. If \code{NULL}, plot all parameters
except those in \code{param_rm}}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If \code{NULL}, nothing is removed}

\item{list_param_names}{Logical. If \code{TRUE}, list the parameter names that will be plot
by this function, but do not actually plot it. This is to help you to use the
\code{param_rm} argument, so you know what the names to be removed can be}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object made from the \code{\link[GGally:ggpairs]{GGally::ggpairs()}} function
}
\description{
Pairs plots for parameters in an MCMC run
}
\details{
Show correlations between parameters as scatterplots in the lower
triangular area, density of each parameter in the diagonals, and the correlation
value with text size scaled to be larger for higher correlations in the upper
triangular area
}
\seealso{
Other MCMC diagnostics plots: 
\code{\link{plot_autocor_mcmc}()},
\code{\link{plot_priors_posts_mcmc}()},
\code{\link{plot_traces_mcmc}()}
}
\concept{MCMC diagnostics plots}
