% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{plot_priors_posts_old}
\alias{plot_priors_posts_old}
\title{Plot priors and posteriors for the given isam model}
\usage{
plot_priors_posts_old(model, priors_only = TRUE)
}
\arguments{
\item{model}{An iscam model object}

\item{priors.only}{Plot the priors only? TRUE/FALSE}
}
\value{
Nothing
}
\description{
Plot priors and posteriors for the given isam model
}
\details{
Plots the priors overlaid on the posteriors for the given iscm model.
  If priors.only is TRUE, only the priors will be plotted.
  The values in the control file (model$ctl$params) for each
  prior are:
  1. ival  = initial value
  2. lb    = lower bound
  3. ub    = upper bound
  4. phz   = ADMB phase
  5. prior = prior distribution funnction
         0 = Uniform
         1 = normal    (p1=mu,p2=sig)
         2 = lognormal (p1=log(mu),p2=sig)
         3 = beta      (p1=alpha,p2=beta)
         4 = gamma     (p1=alpha,p2=beta)
  6. p1 (defined by 5 above)
  7. p2 (defined by 5 above)
}
