% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rdevs-mcmc.R
\name{plot_rdevs_mcmc}
\alias{plot_rdevs_mcmc}
\title{Plot MCMC recruitment deviations for iSCAM models}
\usage{
plot_rdevs_mcmc(
  models,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  palette = iscam_palette,
  base_color = "black",
  x_space = 0.5,
  append_base_txt = NULL,
  ind_letter = NULL,
  leg_loc = c(1, 1),
  probs = c(0.025, 0.5, 0.975),
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{legend_title}{Title for legend}

\item{xlim}{The x limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{palette}{A palette value that is in \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}}

\item{base_color}{A color to prepend to the brewer colors which are set by
\code{palette}. This is called \code{base_color} because it is likely to be a base
model}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\description{
Plot the MCMC recruitment deviation time series trajectories with credible intervals
for iscam models.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_biomass_proj_mcmc}()},
\code{\link{plot_catch_fit_mcmc}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
