% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ref-points.R
\name{plot_ref_points}
\alias{plot_ref_points}
\title{Plot reference point boxplots}
\usage{
plot_ref_points(
  model,
  ref_pts = c("sbo", "bmsy", "msy1", "msy2", "umsy1", "umsy2"),
  ref_desc = c(expression(italic(SB)[0]), expression(italic(B)[MSY]),
    expression(italic(MSY)[1]), expression(italic(MSY)[2]),
    expression(italic(U[MSY])[1]), expression(italic(U[MSY])[2])),
  color = "black",
  fill = "steelblue"
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{ref_pts}{A vector of the names of reference points to plot}

\item{ref_desc}{A vector of fancy descriptions for the values in \code{ref_pts}}

\item{color}{The histogram color (outline of bars)}

\item{fill}{The histogram bar fill color}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} box plots
}
\description{
Plot reference point boxplots
}
