% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ref-points-dist-mcmc.R
\name{plot_ref_points_dist_mcmc}
\alias{plot_ref_points_dist_mcmc}
\title{Plot DFO-style reference point figure with horizontal boxplots and zones
bounded by B0 reference points}
\usage{
plot_ref_points_dist_mcmc(
  model,
  biomass_col = "B2023",
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  proj_catch_vals = NULL,
  line_thickness = 0.75
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{biomass_col}{The column name that you want to shoe output for.
This column will be divided by the SB0 value to get a depletion for
the year in the name of the column}

\item{probs}{Probability values to use in the plot}

\item{proj_catch_vals}{The catch values to include in the plot. If \code{NULL},
include all values present in the list from the output
(\code{model$mcmccalcs$proj_sbt_quants}). If non-null, is a vector of values
of catch to filter the projection biomass table on. Only show those values
in the plot}

\item{line_thickness}{Thickness of reference point lines (vertical lines)}
}
\description{
Plot DFO-style reference point figure with horizontal boxplots and zones
bounded by B0 reference points
}
