% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-selectivity-mcmc.R,
%   R/figures-selectivity-mpd.R
\name{plot_selex_mcmc}
\alias{plot_selex_mcmc}
\alias{plot_selex_mpd}
\title{Plot the selectivity for any combination of gears in an iscam model}
\usage{
plot_selex_mcmc(
  model,
  gear = NULL,
  probs = c(0.025, 0.5, 0.975),
  show_maturity = FALSE,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_alpha = 0.3
)

plot_selex_mpd(model, gear = NULL, show_maturity = FALSE, title = NULL)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{show_maturity}{If `TRUE`, overlay the maturity ogive on the selectivity plots}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See `linetype` in [ggplot2]. Only used if `ci_type` is "line" or "both"}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons. Only used if
`ci_type` is "ribbon" or "both"}

\item{title}{The plot title. NULL means no title}
}
\value{
A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object
}
\description{
Plot the selectivity for any combination of gears in an iscam model

Plot the selectivity for all gears in the iscam model
}
\concept{Selectivity plotting functions}
