% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-mpd.R
\name{plot_selex_mpd}
\alias{plot_selex_mpd}
\title{Plot MPD selectivities for iSCAM models}
\usage{
plot_selex_mpd(model, gear = NULL, show_maturity = FALSE, title = NULL)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{Gear numbers to use in the plot}

\item{show_maturity}{If \code{TRUE}, overlay the maturity ogive on the selectivity plots}

\item{title}{The plot title. NULL means no title}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the selectivity for all gears in the iscam model
}
\seealso{
Other Selectivity plotting functions: 
\code{\link{plot_selex_gear_mcmc}()},
\code{\link{plot_selex_mcmc}()},
\code{\link{plot_tv_selex_mpd}()}
}
\concept{Selectivity plotting functions}
