% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-mcmc-traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot parameter values from each posterior for the estimated parameters}
\usage{
plot_traces(model, plot_sel = FALSE, param_rm = NULL)
}
\arguments{
\item{model}{An iscam model}

\item{plot_sel}{Logical. If `TRUE`, plot only the selectivity parameters.
If `FALSE`, plot only non-selectivity parameters}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If `NULL`, nothing is removed}
}
\value{
A [ggplot2::gglot()] object
}
\description{
Plot parameter values from each posterior for the estimated parameters
}
