% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-autocor-mcmc.R, R/figures-pairs-mcmc.R,
%   R/figures-priors-posts-mcmc.R, R/figures-traces-mcmc.R
\name{plot_autocor}
\alias{plot_autocor}
\alias{plot_pairs_mcmc}
\alias{plot_priors_posts_mcmc}
\alias{plot_traces_mcmc}
\title{Autocorrelation plots for parameters in an MCMC run}
\usage{
plot_autocor(
  model,
  plot_sel = NULL,
  param_rm = c("rho", "vartheta"),
  lag = 100,
  list_param_names = FALSE,
  ...
)

plot_pairs_mcmc(
  model,
  plot_sel = NULL,
  param_rm = c("rho", "vartheta"),
  list_param_names = FALSE
)

plot_priors_posts_mcmc(
  model,
  param_rm = c("sel", "bo", "vartheta", "tau", "sigma", "bmsy"),
  priors_only = FALSE,
  ...
)

plot_traces_mcmc(
  model,
  plot_sel = FALSE,
  param_rm = c("rho", "vartheta"),
  list_param_names = FALSE,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{plot_sel}{Logical. If `TRUE`, plot only the selectivity parameters.
If `FALSE`, plot only non-selectivity parameters. If `NULL`, plot all parameters
except those in `param_rm`}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If `NULL`, nothing is removed}

\item{list_param_names}{Logical. If `TRUE`, list the parameter names that will be plot
by this function, but do not actually plot it. This is to help you to use the
`param_rm` argument, so you know what the names to be removed can be}

\item{...}{Additional arguments passed to [cowplot::plot_grid()]}

\item{priors_only}{Logical. If `TRUE`, plot the priors only. If `FALSE`,
plot both priors and associated posterior histograms}
}
\value{
A [ggplot2::ggplot()] object made from the [forecast::ggAcf()] function

A [ggplot2::ggplot()] object made from the [GGally::ggpairs()] function

Nothing

A [ggplot2::gglot()] object
}
\description{
Autocorrelation plots for parameters in an MCMC run

Pairs plots for parameters in an MCMC run

Priors and posteriors comparison plots for parameters in an MCMC run

Trace plots for parameters in an MCMC run
}
\details{
Show correlations between parameters as scatterplots in the lower
triangular area, density of each parameter in the diagonals, and the correlation
value with text size scaled to be larger for higher correlations in the upper
triangular area

Plots the priors overlaid on the posteriors for the given iscam model.
  The values in the control file (`model$ctl$params`) for each
  prior are:
  1. ival  = initial value
  2. lb    = lower bound
  3. ub    = upper bound
  4. phz   = ADMB phase
  5. prior = prior distribution funnction
         0 = Uniform
         1 = normal    (p1 = mu, p2 = sig)
         2 = lognormal (p1 = log(mu), p2 = sig)
         3 = beta      (p1 = alpha, p2 = beta)
         4 = gamma     (p1 = alpha, p2 = beta)
  6. p1 (defined by 5 above)
  7. p2 (defined by 5 above)

See [plot_fun()] for prior plot details including an documentation on the
vertical lines
}
\concept{MCMC diagnostics plots}
