% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{plot_ts_mpd}
\alias{plot_ts_mpd}
\title{Plot biomass for MPD models against each other. Typically used for bridge models}
\usage{
plot_ts_mpd(
  models,
  model_names = NULL,
  type = "sbt",
  rel = FALSE,
  legend_title = "Bridge model"
)
}
\arguments{
\item{models}{Model list as output by [model_setup()]}

\item{model_names}{A vector of model names to show in plots of the same length as `model`}

\item{type}{Which value to plot. sbt = spawning biomass (and sbo), rt = recruitment (and ro)}

\item{rel}{If `TRUE`, plot relative to initial value}

\item{legend_title}{Title to use for the legend}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot biomass for MPD models against each other. Typically used for bridge models
}
\examples{
\dontrun{
library(here)
library(gfiscamutils)
bridge_models_text <- c("2015 Base model",
                        "Update data to 2019",
                        "Remove HS MSA survey",
                        "Add Discard CPUE index",
                        "Convert model to split sex",
                        "Change fishing year to Feb 21 - Feb 20")
bridge_models_text <- factor(bridge_models_text, levels = bridge_models_text)
drs <- arrowtooth::set_dirs(base_model_dir = "base",
                            bridge_models_dirs = bridge_models_dirs,
                            sens_models_dirs = NULL)
models <- arrowtooth::model_setup(main_dirs = drs,
                                  bridge_models_text = bridge_models_text,
                                  overwrite_rds_files = TRUE)
plot_biomass_mpd(models$bridge_models, bridge_models_text)
}
}
