% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-selectivity.R
\name{plot_tv_selex}
\alias{plot_tv_selex}
\title{Plot selectivities for a gear which has time-varying selectivity blocks (MPD)}
\usage{
plot_tv_selex(model, gear = NULL, title = NULL)
}
\arguments{
\item{model}{An iscam model object as output by [arrowtooth::model_setup()]}

\item{gear}{The gear number}

\item{title}{The plot title. NULL means no title}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot selectivities for a gear which has time-varying selectivity blocks (MPD)
}
