% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-selectivity-tv-mcmc.R
\name{plot_tv_selex_mcmc}
\alias{plot_tv_selex_mcmc}
\title{Plot the (possibly time-varying) selectivity for a prticular gear in an iscam model}
\usage{
plot_tv_selex_mcmc(
  model,
  gear = 1,
  probs = c(0.025, 0.5, 0.975),
  show_maturity = FALSE,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_alpha = 0.3
)
}
\arguments{
\item{model}{An iscam model object as output by [arrowtooth::model_setup()]}

\item{gear}{The gear number in the model}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{show_maturity}{If `TRUE`, overlay the maturity ogive on the selectivity plots}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See `linetype` in [ggplot2]. Only used if `ci_type` is "line" or "both"}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons. Only used if
`ci_type` is "ribbon" or "both"}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the (possibly time-varying) selectivity for a prticular gear in an iscam model
}
