% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-gear-mpd.R
\name{plot_tv_selex_mpd}
\alias{plot_tv_selex_mpd}
\title{Plor MCMC selectivities for iSCAM models}
\usage{
plot_tv_selex_mpd(model, gear = NULL, title = NULL)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot selectivities for a gear which has time-varying selectivity blocks (MPD)
}
\seealso{
Other Selectivity plotting functions: 
\code{\link{plot_selex_gear_mcmc}()},
\code{\link{plot_selex_mcmc}()},
\code{\link{plot_selex_mpd}()}
}
\concept{Selectivity plotting functions}
