% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-vuln-biomass-mcmc.R
\name{plot_vuln_mcmc}
\alias{plot_vuln_mcmc}
\title{Plot MCMC vulnerable biomass for iSCAM models}
\usage{
plot_vuln_mcmc(...)
}
\description{
Plot vulnerable biomass time series plots for MCMC output with credible
intervals
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()}
}
\concept{Time series plotting functions}
