% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{rdata.file}
\alias{rdata.file}
\title{Name of the model rdata file}
\format{
An object of class \code{character} of length 1.
}
\usage{
rdata.file
}
\description{
Name of the model rdata file
}
\keyword{datasets}
