% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read.control.file}
\alias{read.control.file}
\title{Read in the iscam control file}
\usage{
read.control.file(file = NULL, num.gears = NULL,
  num.age.gears = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Filename}

\item{num.gears}{The total number of gears in the datafile}

\item{num.age.gears}{The number of gears with age composition information
in the datafile}

\item{verbose}{Say more}
}
\value{
A list representing the contents on the iscam control file
}
\description{
Read in the iscam control file
}
