% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read.mcmc}
\alias{read.mcmc}
\title{Read in the MCMC results from an iscam model}
\usage{
read.mcmc(model.dir = NULL, verbose = TRUE)
}
\arguments{
\item{model.dir}{Directory in which the model output resides}

\item{verbose}{Say more}
}
\value{
A list representing the MCMC output of the iscam model,
  or NULL if there was a problem or there werer no MCMC output files
}
\description{
Read in the MCMC results from an iscam model
}
