% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read.par.file}
\alias{read.par.file}
\title{Read in the contents of the iscam par file}
\usage{
read.par.file(file = NULL)
}
\arguments{
\item{file}{Filename}
}
\value{
A list representing the contents of the iscam par file
}
\description{
Read in the contents of the iscam par file
}
