% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read.projection.file}
\alias{read.projection.file}
\title{Read in the contents of the iscam projection file}
\usage{
read.projection.file(file = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Filename}

\item{verbose}{Say more}
}
\value{
A list representing the contents of the iscam projection file
}
\description{
Read in the contents of the iscam projection file
}
