% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-file.R
\name{read_data_file}
\alias{read_data_file}
\title{Read in the iscam data file}
\usage{
read_data_file(file = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Filename (full path)}

\item{verbose}{Say more}
}
\value{
A list representing the contents of the iscam data file
}
\description{
Read in the iscam data file
}
\details{
This file was checked out back in time because it had been
changed to work on Petrale and had to be changed back to work on Arrowtooth.
The command run was:
git checkout 815275c~1 -- R/read-data-file.R
The command to bring it back should be:
git checkout 424f727 -- R/read-data-file.R
}
