% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read_mcmc}
\alias{read_mcmc}
\title{Read in the MCMC results from an iscam model, reformat them, and
remove some records as defined by \code{burnin} and \code{thin}}
\usage{
read_mcmc(model, burnin = 1000, thin = 1, load_proj = TRUE, ...)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{burnin}{The number of MCMC records to remove for burnin period}

\item{thin}{Remove every nth record for thinning of MCMC output}

\item{load_proj}{Load projection output}

\item{...}{Extra arguments}
}
\value{
A list of data frames representing the MCMC output of the
iscam model, or NULL if there was a problem or there were no MCMC
output files
}
\description{
Read in the MCMC results from an iscam model, reformat them, and
remove some records as defined by \code{burnin} and \code{thin}
}
