% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read_par_file}
\alias{read_par_file}
\title{Read in the contents of the iscam par file}
\usage{
read_par_file(file = NULL)
}
\arguments{
\item{file}{Filename}
}
\value{
A list representing the contents of the iscam par file
}
\description{
Read in the contents of the iscam par file
}
