% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-projection-file.R
\name{read_projection_file}
\alias{read_projection_file}
\title{Read in the contents of the iscam projection file}
\usage{
read_projection_file(file = NULL)
}
\arguments{
\item{file}{Filename}
}
\value{
A list representing the contents of the iscam projection file
}
\description{
Read in the contents of the iscam projection file
}
