% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_psv}
\alias{read_psv}
\title{Read ADMB-generated psv file}
\usage{
read_psv(fn, n_samples = 10000)
}
\arguments{
\item{fn}{Filename}

\item{nsamples}{Number of samples}
}
\description{
Read ADMB-generated psv file
}
\details{
The psv file is generated when the MCMC is run in ADMB and is used to
perform the calculations needed for the -mceval step. It contains the
posterior values in binary fowmat.
}
