% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_gear_iscam.R
\name{rename_gear_iscam}
\alias{rename_gear_iscam}
\title{Read in an ISCAM data file, replace all instances of the old name
with the new name and save the file, overwriting it}
\usage{
rename_gear_iscam(fn, old_gear_name = NULL, new_gear_name = NULL)
}
\arguments{
\item{fn}{The iSCAM data file}

\item{old_gear_name}{The old name or abbreviation for the gear/fleet}

\item{new_gear_name}{The new name or abbreviation for the gear/fleet}
}
\value{
Nothing
}
\description{
Read in an ISCAM data file, replace all instances of the old name
with the new name and save the file, overwriting it
}
