% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-gear-names.R
\name{replace_gear_name_and_save}
\alias{replace_gear_name_and_save}
\title{Replace gear names in objects in an iSCAM output list, from an RDS file
and re-save the file}
\usage{
replace_gear_name_and_save(fn, ...)
}
\arguments{
\item{fn}{The RDS file to modify}

\item{...}{Arguments passed to \code{\link[=replace_gear_names]{replace_gear_names()}}}
}
\value{
Nothing
}
\description{
Replace gear names in objects in an iSCAM output list, from an RDS file
and re-save the file
}
