% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-gear-names.R
\name{replace_gear_names}
\alias{replace_gear_names}
\title{Change the gear names in already-loaded models}
\usage{
replace_gear_names(
  models,
  old_gear_names = NULL,
  new_gear_names = NULL,
  old_gear_abbrevs = NULL,
  new_gear_abbrevs = NULL
)
}
\arguments{
\item{models}{A list of outputs of \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{old_gear_names}{A vector of the old gear names to replace.
See model object (\code{model$dat$gear_names}) to see what they currently are}

\item{new_gear_names}{A vector of the new gear names. Must be the same
length as \code{old_gear_names}}

\item{old_gear_abbrevs}{A vector of the old gear abbreviations to replace.
See model object (\code{model$dat$gear_abbrevs}) to see what they currently are}

\item{new_gear_abbrevs}{A vector of the new gear abbrevs. Must be the same
length as \code{old_gear_abbrevs}}
}
\value{
A modified list of \code{models}
}
\description{
Change the gear names in already-loaded models. This function takes
a list of outputs from \code{\link[=load_iscam_files]{load_iscam_files()}} and changes the gear names
and/or abbreviations inside them
}
