% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{retro.dir}
\alias{retro.dir}
\title{Name of the retrospectives directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
retro.dir
}
\description{
Name of the retrospectives directory
}
\keyword{datasets}
