% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{roll_mean_na}
\alias{roll_mean_na}
\title{Fill in NA values with rolling mean of previous values}
\usage{
roll_mean_na(dat, n, omit_na = TRUE)
}
\arguments{
\item{dat}{Data frame}

\item{n}{N value for minimum}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\description{
Update the NAs in a vector with the mean of the previous values. The number
of values used can be less than n for NAs near the start of the vector, and
will be a maximum of n for values further along the vector. The value will
remain NA if no non-NA values are available.
}
