% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-iscam.R
\name{run_iscam}
\alias{run_iscam}
\title{Run an iscam model through the [base::shell()] command}
\usage{
run_iscam(
  model_dir,
  mcmc_mode = FALSE,
  mcmc_dir = "mcmc",
  mcmc_chain_length = 500,
  mcsave = 1,
  mcscale = 0,
  maxfn = 2000,
  crit = 1e-04,
  nox = TRUE,
  ...
)
}
\arguments{
\item{model_dir}{The directory in which the iscam model lies}

\item{mcmc_mode}{Logical. TRUE means enter the `mcmc_dir` and run an
MCMC with the remaining arguments}

\item{mcmc_dir}{Directory to run the MCMC in}

\item{mcmc_chain_length}{MCMC chain length}

\item{mcsave}{MCMC save every `mcsave`th posterior}

\item{mcscale}{Scale step size for the first `mcscale` evaluations}

\item{maxfn}{Maximum number of function evaluations}

\item{crit}{Gradient convergence criterion}

\item{nox}{Logical. If TRUE, do not show gradient info to the screen}

\item{...}{Passed on to other functions}
}
\value{
[base::invisible()]
}
\description{
Run an iscam model through the [base::shell()] command
}
