% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-iscam.R
\name{run_multiple_iscam}
\alias{run_multiple_iscam}
\title{Run multiple iscam models in parallel}
\usage{
run_multiple_iscam(model_dirs, ...)
}
\arguments{
\item{model_dirs}{A vector of model directories}

\item{...}{Arguments passed to [run_iscam()]}
}
\value{

}
\description{
Run multiple iscam models in parallel
}
