% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Season2Year}
\alias{Season2Year}
\title{Function to transform from Season to Year}
\usage{
Season2Year(dat)
}
\arguments{
\item{dat}{Data frame}
}
\description{
Function to transform from Season to Year
}
