% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-selectivity.R
\name{sel_param_est_mpd_table}
\alias{sel_param_est_mpd_table}
\title{Make a table of selectivity parameter estimate comparisons for MPD models}
\usage{
sel_param_est_mpd_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = NULL,
  ...
)
}
\arguments{
\item{models}{A list of An ISCAM model objects (class \link{mdl_lst_cls})}

\item{digits}{Number of decimal places to show}

\item{french}{If \code{TRUE}, display the table in French}

\item{model_col_widths}{A vector of column widths. If \code{NULL}, the defaults
are used}

\item{...}{Arguments passed to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
a \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of selectivity parameter estimate comparisons for MPD models
}
