% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-dirs.R
\name{set_dirs}
\alias{set_dirs}
\title{Set and return all model directories for the project}
\usage{
set_dirs(
  models_dir = NA,
  nongit_dir = NA,
  base_models_dir = "01-base-models",
  bridge_models_dir = "02-bridge-models",
  sens_models_dir = "03-sens-models",
  retro_models_dir = "04-retrospective-models",
  base_models_dirs = "01-base-model",
  bridge_models_dirs = NA,
  sens_models_dirs = NA,
  retro_models_dirs = NA,
  prepend_to_bridge = NA,
  suppress_warnings = FALSE
)
}
\arguments{
\item{models_dir}{Path in which the model directories are
located for the current assessment year}

\item{nongit_dir}{Path in which the non-git files for the assessment are
located}

\item{base_models_dir}{Name of the base models directory}

\item{bridge_models_dir}{Name of the bridging models directory}

\item{sens_models_dir}{Name of the sensitivity models directory}

\item{base_models_dirs}{A vector of subdirectory names in
\code{base_models_dir} that each contain an individual base model}

\item{bridge_models_dirs}{A vector of subdirectory names in
\code{bridge_models_dir} that each contain an individual bridge model}

\item{sens_models_dirs}{A vector of subdirectory names in \code{sens_models_dir}
that each contain an individual sensitivity model}

\item{suppress_warnings}{If \code{TRUE}, warnings about directories not existing
will not be shown}

\item{retrospectives_models_dir}{Name of the retrospectives models directory}

\item{retrospectives_models_dirs}{A vector of subdirectory names in
\code{retrospectives_models_dir}that each contain an individual retrospective
model}
}
\value{
A list of vectors of directory names, which will have \code{NA} elements
for those which do not exist:
\enumerate{
\item The input directory for \code{models_dir}
\item The input directory for \code{nongit_dir}
\item The input directory for \code{last_yr_base_model_dir}
\item The input directory for \code{base_model_dir}
\item The input directory for \code{bridge_models_dir}
\item The input directory for \code{sens_models_dir}
\item The input directory for \code{retrospective_models_dir}
\item The base model directories
\item The bridge model directories
\item The sensitivity model directories
}
}
\description{
Set and return all model directories for the project
}
\details{
Sensitivity groups will have the base model directory prepended,
The bridge model groups can have \code{last_yr_base_model_dir}
prepended if \code{prepend_to_bridge} is set to \code{TRUE} for the model group.
If directories do not exist, they will be assigned \code{NULL} and a warning
issued. If the \code{models_dir} does not exist, an error will be thrown.
}
