% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-gear-language.R
\name{set_gear_language}
\alias{set_gear_language}
\title{Translate the gear names and abbreviations in the models list for all
models to either French of English}
\usage{
set_gear_language(models, curr_gear_lang = NULL)
}
\arguments{
\item{models}{A list of the models}

\item{curr_gear_lang}{The language that the gear names are currently
stored as.}
}
\value{
A modified \code{models} list
}
\description{
Translate the gear names and abbreviations in the models list for all
models to either French of English
}
\details{
Depends on the return value of fr() which is just the value
of the R global option \code{french}
}
