% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{simplify_names}
\alias{simplify_names}
\title{Simplify MCMC parameter names for this package code}
\usage{
simplify_names(dat)
}
\arguments{
\item{dat}{A data frame of posteriors as seen in the MCMC output csv files}
}
\value{
The original data frame with the area and group prefixes removed from the column names
}
\description{
Simplify MCMC parameter names for this package code
}
