% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split.prior.info}
\alias{split.prior.info}
\title{Get priors information from prior.str}
\usage{
\method{split}{prior.info}(prior.str, dec.points = 1,
  first.to.lower = FALSE)
}
\arguments{
\item{prior.str}{is a string like Lognormal(2.0,1.01)}

\item{dec.points}{number of decimal points to return}

\item{first.to.lower}{makes the first letter of the name of the prior lower case}
}
\value{
a vector of length 3, e.g.:"Lognormal", 2.0, 1.01
}
\description{
Get priors information from prior.str
}
