% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-props.R
\name{standardize_comps}
\alias{standardize_comps}
\title{Update an iSCAM \code{dat} file by standardizing all the age comps}
\usage{
standardize_comps(fns)
}
\arguments{
\item{fns}{A vector of the iSCAM \code{dat} filenames to process}
}
\value{
Nothing
}
\description{
Update an iSCAM \code{dat} file by standardizing all the age comps
}
