% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{strip_areas_groups}
\alias{strip_areas_groups}
\title{Remove the area and group prefixes for the mcmc data}
\usage{
strip_areas_groups(dat)
}
\arguments{
\item{dat}{A data frame of posteriors as seen in the MCMC output csv files}
}
\value{
The original data frame with the area and group prefixes removed
from the column names
}
\description{
Remove the area and group prefixes for the mcmc data
}
\details{
This is a hack function to remove the area and group prefixes for the
  mcmc data 'dat'. The reason is that for now we are just working with a
  single group and area, and the extra text in the parameter names are
  confusing, eg. 'ro_gr1' will become just 'ro'. If you make a model with
  more than one group or area this will need to be revisited. Also, this
  removes 'f' which is assumed to be the objective function value. Note
  that q1, q2, q3... will stay the same and m1 and m2 will remain if the
  model was two-sex.
}
