% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{strip_static_params}
\alias{strip_static_params}
\title{Strip static parameters from MCMC output data}
\usage{
strip_static_params(model, dat)
}
\arguments{
\item{model}{An iSCAM model object as created in [load_iscam_files()]}

\item{dat}{A vector of the parameter names to remove}
}
\value{
The parameters data frame for the model given with static parameters removed
}
\description{
Strip static parameters from MCMC output data
}
\details{
Strip out the static (non-estimated) parameters from the mcmc output data
  for the given scenario. We only need to see estimated parameters on the
  diagnostic plots. If there are no static parameters, NULL will be returned
}
