% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_na}
\alias{sum_na}
\title{Calculate sum if there are non-NA values, return NA if all values are NA}
\usage{
sum_na(x, omit_na = TRUE)
}
\arguments{
\item{x}{value}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\description{
Calculate sum if there are non-NA values, return NA if all values are NA
}
