% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-bio-depl.R
\name{table_bio_depl}
\alias{table_bio_depl}
\title{Create a table containing biomass and depletion estimates}
\usage{
table_bio_depl(
  model,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)
}
\value{
an xtable
}
\description{
Make a table with both spawning biomass and depletion in it.
Based on \code{\link[=table_ts_value]{table_ts_value()}}, but wider with both and extra headers
}
