% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of catch by gear type}
\usage{
table_catch(
  catch_df,
  start_yr = NULL,
  area = NULL,
  gear_col_widths = "5em",
  scale_factor = 1000,
  ...
)
}
\arguments{
\item{catch_df}{A data frame as output by \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}

\item{start_yr}{The year to start the table or \code{NULL} to include all data}

\item{area}{If \code{NULL}, sum together all catch by year for all areas found
in \code{catch_df}. Otherwise, a single area name (eg. 3CD or 5ABCDE). If the
area name is not in \code{catch_df$area}, an error will be thrown}

\item{gear_col_widths}{Width to make the columns for the gear types}

\item{scale_factor}{Value to divide all the catches by to make the table
in tonnes}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of catch by gear type
}
