% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch_area}
\alias{table_catch_area}
\title{Return a table by area, where the areas are defined by the unique values
in the \code{area} column in table \code{catch_df}}
\usage{
table_catch_area(
  catch_df,
  start_yr = NULL,
  gear_col_widths = "5em",
  scale_factor = 1000,
  ...
)
}
\arguments{
\item{catch_df}{A data frame as output by \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}

\item{start_yr}{The year to start the table or \code{NULL} to include all data}

\item{gear_col_widths}{Width to make the columns for the gear types}

\item{scale_factor}{Value to divide all the catches by to make the table
in tonnes}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\description{
Return a table by area, where the areas are defined by the unique values
in the \code{area} column in table \code{catch_df}
}
\keyword{internal}
