% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch_fleet}
\alias{table_catch_fleet}
\title{Return a table by fleet, where the fleets are determined by the \code{...}
argument as data frames}
\usage{
table_catch_fleet(
  catch_df_lst = NULL,
  fleet_nms = NULL,
  start_yr = NULL,
  gear_col_widths = "5em",
  scale_factor = 1000,
  ret_df = FALSE,
  show_total_col = TRUE,
  ...
)
}
\arguments{
\item{catch_df_lst}{A list of data frames with catch, one for each fleet
which have had \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}} applied}

\item{fleet_nms}{A vector of names for the fleets, to appear on the table.
Must be the same length as \code{catch_df_lst}}

\item{start_yr}{The year to start the table or \code{NULL} to include all data}

\item{gear_col_widths}{Width to make the columns for the gear types}

\item{scale_factor}{Value to divide all the catches by to make the table
in tonnes}

\item{ret_df}{If \code{TRUE} return a data frame, if \code{FALSE}, return an
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{show_total_col}{If \code{TRUE}, show a column with the total landings
and discards for all fleets}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\description{
Return a table by fleet, where the fleets are determined by the \code{...}
argument as data frames
}
\keyword{internal}
