% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-decision.R
\name{table_decision}
\alias{table_decision}
\title{Create a decision table with a single reference point or probabilities
of biomass declining or increasing}
\usage{
table_decision(
  model,
  catch_vals = NULL,
  ret_df = FALSE,
  format = c("latex", "html"),
  bold_header = FALSE,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{catch_vals}{The catch values to include in the table. If \code{NULL},
all catch values present in the projections will be used}

\item{ret_df}{Logical. If \code{TRUE} return a data frame with the values,
instead of the \code{\link[csasdown:csas_table]{csasdown::csas_table()}} formatted table}

\item{format}{One of "html" or "latex"}

\item{bold_header}{If \code{TRUE}, make the headers bold. This only works
if the \code{format} is 'html'. If it is 'latex', you must paste latex bold
macros around the column headers manually}

\item{...}{Arguments to be passed to \code{\link[=get_proj_biomass_raw]{get_proj_biomass_raw()}},
\code{\link[=calc_probs_biomass]{calc_probs_biomass()}}, and \code{\link[=csas_table]{csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a decision table with a single reference point or probabilities
of biomass declining or increasing
}
