% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-decisions.R
\name{table_decisions}
\alias{table_decisions}
\title{Create a decision table for iSCAM models}
\usage{
table_decisions(
  model,
  format = c("latex", "html"),
  bo_refpts = c(0.2, 0.35, 0.4),
  num_proj_yrs = 3,
  digits = 2,
  ret_df = FALSE,
  col_widths = NULL,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{format}{See \code{format} in \code{\link[knitr:kable]{knitr::kable()}}}

\item{bo_refpts}{Vector of proportional values for the reference
points}

\item{num_proj_yrs}{The number of projection years to show in the table.
The output may have more but only this many will be shown. If \code{NULL},
all projection years will be shown}

\item{digits}{Number of decimal places for the values in the table}

\item{ret_df}{Logical. If \code{TRUE} return a data frame with the values,
instead of the \code{\link[knitr:kable]{knitr::kable()}} formatted table}

\item{col_widths}{Widths of columns, see \code{\link[kableExtra:column_spec]{kableExtra::column_spec()}}}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Produce a decision table for the given iscam model
}
