% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-iscam-param-desc.R
\name{table_iscam_param_desc}
\alias{table_iscam_param_desc}
\title{Create the ISCAM model parameter description table as found in the
Model Description appendix}
\usage{
table_iscam_param_desc(model, bold_header = TRUE, ...)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{...}{Arguments passed to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[knitr:kable]{knitr::kable()}} table
}
\description{
Create the ISCAM model parameter description table as found in the
Model Description appendix
}
