% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-maturity.R
\name{table_maturity}
\alias{table_maturity}
\title{Table of maturity estimates based on data only}
\usage{
table_maturity(surv_samples, col_widths = NULL, return_df = FALSE, ...)
}
\arguments{
\item{surv_samples}{A \code{survey_samples} list as output by
\code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{col_widths}{Widths for columns, except the Parameter column}

\item{return_df}{Logical. If \code{TRUE}, return the \link{data.frame} and not
the \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{digits}{Number of decimal places for the values in the table}
}
\description{
Table of maturity estimates based on data only
}
