% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est-mcmc.R
\name{table_param_est_mcmc}
\alias{table_param_est_mcmc}
\title{Create a table of parameter estimates for iSCAM models}
\usage{
table_param_est_mcmc(
  model,
  digits = 2,
  probs = c(0.025, 0.5, 0.975),
  model_col_widths = NULL,
  ret_df = FALSE,
  show_year_range = TRUE,
  bold_headers = TRUE,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{digits}{Number of decimal places for the values in the table}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{ret_df}{If \code{TRUE}, return a data frame. If \code{FALSE} return a
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{show_year_range}{Logical. If \code{TRUE}, show the "Year range" column
in the output table}

\item{bold_headers}{If \code{TRUE}, make all column headers bold}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table parameter estimates and priors for iSCAM models
}
