% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-est-mpd.R
\name{table_param_est_mpd}
\alias{table_param_est_mpd}
\title{Create a table of MPD parameter comparisons}
\usage{
table_param_est_mpd(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = "5em",
  ...
)
}
\arguments{
\item{models}{A list of iSCAM models as output from \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{digits}{Number of digits to show}

\item{french}{If \code{TRUE} translate to French}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
An \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of parameter estimate comparisons for MPD models.
Also show B0, Fmsy. Bmsy, and msy
}
