% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-phi-mpd.R
\name{table_param_phi_mpd}
\alias{table_param_phi_mpd}
\title{Create a table comparing phi parameter}
\usage{
table_param_phi_mpd(
  model,
  format = "latex",
  yrs = NULL,
  digits = 2,
  col_widths = "5em",
  ret_df = FALSE,
  by_sex = FALSE,
  ...
)
}
\arguments{
\item{model}{An iSCAM model as output from \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{format}{See argument in \code{\link[knitr:kable]{knitr::kable()}}}

\item{yrs}{A vector of years to include in the table. If \code{NULL}, all
will be included}

\item{digits}{Number of digits to show}

\item{col_widths}{Widths for columns, except the Parameter column}

\item{ret_df}{If \code{TRUE}, return a data.frame instead of the
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{by_sex}{If \code{TRUE}, the table will have two rows per year, one for
each sex. If \code{FALSE}, the female value will be used}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
An \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of comparisons of estimates of the phi parameter for the
Dirichlet Multinomial for MPD models.
}
