% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-settings.R
\name{table_param_settings}
\alias{table_param_settings}
\title{Create a table summarizing parameterization of an iSCAM model}
\usage{
table_param_settings(model, french = FALSE, ...)
}
\arguments{
\item{model}{A single output model}

\item{french}{If \code{TRUE} translate to French}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\description{
Create a table showing number of parameters estimated and prior
parameterizations
}
