% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-params-est.R
\name{table_params_est}
\alias{table_params_est}
\title{Create a table parameter estimates and priors for iSCAM models}
\usage{
table_params_est(
  model,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)
}
\value{
an xtable
}
\description{
Create a table parameter estimates and priors for iSCAM models
}
