% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ref-points-mcmc.R
\name{table_ref_points_mcmc}
\alias{table_ref_points_mcmc}
\title{Make a table of reference points for MCMC iSCAM models}
\usage{
table_ref_points_mcmc(
  models,
  type = c("median", "ci"),
  digits = 2,
  probs = c(0.025, 0.5, 0.975),
  model_col_widths = "5em",
  ...
)
}
\arguments{
\item{type}{Use 'median' to show median parameter values or 'ci' for a
dash-separated range for the credible interval}

\item{digits}{Number of decimal places for the values in the table}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of reference points for MCMC iSCAM models
}
