% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sens-params.R
\name{table_sens_param_changes}
\alias{table_sens_param_changes}
\title{Create a table summarizing changes for sensitivity models}
\usage{
table_sens_param_changes(sens_desc_vec, sens_change_vec)
}
\value{
an xtable
}
\description{
Create a table summarizing changes for sensitivity models
}
