% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sens-q.R
\name{table_sens_q}
\alias{table_sens_q}
\title{Create a table comparing estimates of q across models}
\usage{
table_sens_q(
  models,
  model.names,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\value{
an xtable
}
\description{
Make a table of the values of q, including quantiles
}
