% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ts-values.R
\name{table_ts_values}
\alias{table_ts_values}
\title{Create a table of time series estimates}
\usage{
table_ts_values(
  model,
  type,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)
}
\value{
an xtable
}
\description{
Make a table for values such as biomass and recruitment
}
