% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ts-values.R
\name{table_ts_values_mcmc}
\alias{table_ts_values_mcmc}
\title{Create a table of time series estimates for MCMC iSCAM models}
\usage{
table_ts_values_mcmc(
  models,
  value = c("sbt", "rt", "ft", "ut", "depl"),
  type = c("median", "ci"),
  start_yr = NULL,
  digits = 2,
  probs = c(0.025, 0.5, 0.975),
  model_col_widths = "5em",
  ...
)
}
\arguments{
\item{value}{An output value to produce the table for. Can be
'sbt' (spawning biomass), 'rt' (recruitment), 'ft' (fishing mortality),
'ut' (exploitation rate), or 'depl' (depletion)}

\item{type}{Use 'median' to show median parameter values or 'ci' for a
dash-separated range for the credible interval}

\item{start_yr}{Year to start the table with. If \code{NULL}, all years will
be included}

\item{digits}{Number of decimal places for the values in the table}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of medians and/or credible intervals for values such as
biomass, recruitment, depletion, and fishing mortality
}
