% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tex}
\alias{tex}
\title{Create an expression from a string that may have a latex string in it}
\usage{
tex(text)
}
\arguments{
\item{text}{The string}
}
\value{
An object with classes \code{latexexpression} and \code{expression}. See
\code{\link[latex2exp:TeX]{latex2exp::TeX()}}
}
\description{
Create an expression from a string that may have a latex string in it
}
