% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unique_na}
\alias{unique_na}
\title{Calculate unique if there are non-NA values, return NA if all values are NA}
\usage{
unique_na(x)
}
\arguments{
\item{x}{Value}
}
\description{
Calculate unique if there are non-NA values, return NA if all values are NA
}
