% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vec2df}
\alias{vec2df}
\title{Create a one-row tibble from a vector of values}
\usage{
vec2df(vec, nms = NULL)
}
\arguments{
\item{vec}{A vector}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row
}
\description{
Create a one-row tibble from a vector of values
}
