% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{verify_models}
\alias{verify_models}
\title{Verify that models object is a list of iscam models and that models is the same length as models_names}
\usage{
verify_models(models, models_names)
}
\arguments{
\item{models}{a list of iscam models}

\item{models_names}{a vector of names for the models}
}
\description{
Verify that models object is a list of iscam models and that models is the same length as models_names
}
