% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_projection_file.R
\name{write_projection_file}
\alias{write_projection_file}
\title{Write an ISCAM projection file (pfc file)}
\usage{
write_projection_file(fn, proj_lst, overwrite = FALSE)
}
\arguments{
\item{fn}{The filename}

\item{proj_lst}{A list with the same structure as that returned by
\code{\link[=read_projection_file]{read_projection_file()}}}

\item{overwrite}{If \code{TRUE}, overwrite the file if it exists}
}
\value{
Nothing
}
\description{
Write an ISCAM projection file (pfc file)
}
