% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wt_mean_na}
\alias{wt_mean_na}
\title{Calculate weighted mean if there are non-NA values, return NA if all values
are NA}
\usage{
wt_mean_na(x, w, omit_na = TRUE)
}
\arguments{
\item{x}{value}

\item{w}{See [weighted.mean()]}

\item{omit_na}{Omit NA TRUE or FALSE}
}
\value{
This version retuns NA if x is all NA, otherwise it returns the weighted mean
}
\description{
Calculate weighted mean if there are non-NA values, return NA if all values
are NA
}
