% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{biomass.plot.mpd}
\alias{biomass.plot.mpd}
\title{Plot the MPD biomass time series along with the associated retrospectives}
\usage{
biomass.plot.mpd(model, depl = FALSE, xlim = NA, offset = 0.7,
  leg = TRUE)
}
\arguments{
\item{model}{an iscam model object}

\item{depl}{if TRUE, plot the depletion line(s)}

\item{offset}{horizontal offset for B0 points in depletion plot}

\item{leg}{show the legend? Logical}
}
\value{
a ggplot object
}
\description{
Plot the MPD biomass time series along with the associated retrospectives
}
