% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{biomass.plot.mpd}
\alias{biomass.plot.mpd}
\title{Plot the MPD biomass time series along with the associated retrospectives}
\usage{
biomass.plot.mpd(model, depl = FALSE, offset = 0.7, leg = TRUE)
}
\arguments{
\item{model}{an iscam model object}

\item{offset}{horizontal offset for B0 points in depletion plot}

\item{leg}{show the legend? Logical}

\item{show.bo.line}{Show the B0-based reference point lines}
}
\value{
a ggplot object
}
\description{
Plot the MPD biomass time series along with the associated retrospectives
}
