% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc.mcmc}
\alias{calc.mcmc}
\title{Perform some quantile calculations on the MCMC posteriors}
\usage{
calc.mcmc(model, burnin = 1000, thin = 1, lower = 0.025,
  upper = 0.975, load.proj = TRUE, ...)
}
\arguments{
\item{model}{An iscam model object}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{thin}{The thinning to apply to the MCMC posterior samples}

\item{lower}{Lower quantile value to apply to MCMC samples}

\item{upper}{Upper quantile value to apply to MCMC samples}

\item{load.proj}{Load the projections from the MCMC and do the calculations}

\item{...}{arguments to pass to [calc.probabilities()]
to construct the decision tables}
}
\value{
A list of each parameter for which quantiles were calculated
}
\description{
Perform some quantile calculations on the MCMC posteriors
}
