% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mcmc.R
\name{calc_mcmc}
\alias{calc_mcmc}
\title{Perform some quantile calculations on the MCMC posteriors}
\usage{
calc_mcmc(
  model,
  burnin = 1000,
  thin = 1,
  probs = c(0.025, 0.5, 0.975),
  index_scale = 1e+06,
  load_proj = TRUE,
  ...
)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{burnin}{The number of MCMC records to remove for burnin period}

\item{thin}{Remove every nth record for thinning of MCMC output}

\item{probs}{The probabilities to use for \code{quantile()} calculations}

\item{index_scale}{Number to multiply the index values by so they match the
\code{survey_index} values}

\item{load_proj}{Do the calculations for the projections}

\item{...}{arguments to pass to \code{\link[=calc_proj_probs]{calc_proj_probs()}} to construct the
decision tables}
}
\value{
A list of each parameter for which quantiles were calculated
}
\description{
Perform some quantile calculations on the MCMC posteriors
}
