% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{calc_proj_probs}
\alias{calc_proj_probs}
\title{Extract and calculate probabilities from the iscam projection model}
\usage{
calc_proj_probs(model, burnin, thin, fixed_cutoffs, ...)
}
\arguments{
\item{model}{An iscam model object}

\item{burnin}{The number of samples to burn away from the beginning of the MCMC}

\item{thin}{The thinning to apply to the MCMC posterior samples}

\item{fixed_cutoffs}{A vector of catch cutoffs to use in decision tables}

\item{...}{Extra arguments}
}
\value{
A data frame which has its names formatted for latex
}
\description{
Extract and calculate probabilities from the iscam projection model
}
\details{
Extract and calculate probabilities from the projection model.
  Used for decision tables in the document (see make.decision.table())
  in tables-decisions.r
}
