% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_sig_tau}
\alias{calc_sig_tau}
\title{Calculation of sigma and tau from rho and vartheta}
\usage{
calc_sig_tau(rho, vartheta)
}
\arguments{
\item{rho}{parameter rho from iscam model}

\item{vartheta}{parameter vartheta from iscam model}
}
\value{
a list of length 2, the calculated tau and sigma parameters
}
\description{
Calculation of sigma and tau from rho and vartheta
}
\details{
Total variance is given by ϑ (`vartheta`) and the proportion of the total
variance that is associated with observation errors is given by ρ (`rho`),
the variance is partitioned into observation errors (σ^2) (`sigma`^2) and
process errors (τ^2) (`tau`^2).
}
