% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{create.rdata.file}
\alias{create.rdata.file}
\title{Create a .RData file to hold the model's data and outputs}
\usage{
create.rdata.file(model.dir, ovwrt.rdata = FALSE, ...)
}
\arguments{
\item{model.dir}{Directory name for all models location}

\item{ovwrt.rdata}{overwrite the RData file if it exists? TRUE/FALSE}

\item{...}{arguments to pass to [load.iscam.files()]}
}
\value{
Nothing
}
\description{
Create a .RData file to hold the model's data and outputs
}
\details{
If an RData file exists, and overwrite is FALSE, return immediately.
  If no RData file exists, the model will be loaded from outputs into
  an R list and saved as a .RData file in the correct directory.
  When this function exits, a .RData file will be located in the
  directory given by model.name.
  Assumes the files model-setup.r and utilities.r has been sourced.
}
