% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{delete.rdata.files}
\alias{delete.rdata.files}
\title{Delete .Rdata files from all subdirectories}
\usage{
delete.rdata.files(del.dir = model.dir)
}
\arguments{
\item{del.dir}{The directory which has subdirectories
from which the files will be deleted}
}
\value{
Nothing
}
\description{
Delete .Rdata files from all subdirectories
}
