% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{dos.append.to.log}
\alias{dos.append.to.log}
\title{string to append to a program shell call to append the output of stdout and stderr
to a file in DOS}
\format{
An object of class \code{character} of length 1.
}
\usage{
dos.append.to.log
}
\description{
character
}
\keyword{datasets}
