% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R, R/gfiscamutils.R
\docType{package}
\name{gfiscamutils}
\alias{gfiscamutils}
\title{gfiscamutils: A package for loading gfiscam model output, with plotting
  and table-making functions}
\description{
See the README on
\href{https://github.com/pbs-assess/gfiscamutils#readme}{GitHub}
}
