% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{package}
\name{gfiscamutils}
\alias{gfiscamutils}
\alias{gfiscamutils-package}
\title{gfiscamutils: A package for loading gfiscam model output, with plotting
  and table-making functions}
\description{
gfiscamutils: A package for loading gfiscam model output, with plotting
  and table-making functions
}
