% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{linux.pipe.to.log}
\alias{linux.pipe.to.log}
\title{string to append to a program shell call to pipe the output of stdout and stderr
to a file in Linux}
\format{An object of class \code{character} of length 1.}
\usage{
linux.pipe.to.log
}
\description{
character
}
\keyword{datasets}
