% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load.models}
\alias{load.models}
\title{Load the iscam models and return as a list of iscam model objects}
\usage{
load.models(model.dir.names)
}
\arguments{
\item{model.dir.names}{Vector of directory names of models to be loaded}
}
\value{
A list of iscam model objects
}
\description{
Load the iscam models and return as a list of iscam model objects
}
