% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{log.file}
\alias{log.file}
\title{file name for log file piped to using shell call to iscam program}
\format{An object of class \code{character} of length 1.}
\usage{
log.file
}
\description{
character
}
\keyword{datasets}
