% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.autocor.plot}
\alias{make.autocor.plot}
\title{Plot the autocorrelation of estimated parameters}
\usage{
make.autocor.plot(model, mcmc_burnin = 1000)
}
\arguments{
\item{model}{An iscam model object}
}
\value{
Nothing
}
\description{
Plot the autocorrelation of estimated parameters
}
