% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-mcmc-diagnostics.R
\name{make.autocor.plot}
\alias{make.autocor.plot}
\title{Plot the autocorrelation of estimated parameters}
\usage{
make.autocor.plot(model, mcmc_burnin = 1000, lag_max)
}
\arguments{
\item{model}{An iscam model object}

\item{mcmc_burnin}{Number for burnin}

\item{lag_max}{Maximum value at which to calculate acf}
}
\value{
Nothing
}
\description{
Plot the autocorrelation of estimated parameters
}
