% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R
\name{make.parameters.est.table}
\alias{make.parameters.est.table}
\alias{make.ref.points.table}
\alias{make.value.table}
\alias{make.biomass.depletion.table}
\alias{make.sens.parameter.table}
\alias{make.sens.q.table}
\alias{make.variance.table}
\alias{make.catchability.parameters.table.q.sens}
\title{Make a table of parameter estimates and priors}
\usage{
make.parameters.est.table(
  model,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)

make.ref.points.table(
  model.am2,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)

make.value.table(
  model,
  type,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)

make.biomass.depletion.table(
  model,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)

make.sens.parameter.table(
  tab,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.sens.q.table(
  models,
  model.names,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.variance.table(
  var,
  which.model = 1,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.catchability.parameters.table.q.sens(
  qa.lst,
  qb.lst,
  qc.lst,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{translate}{Logical. Translate to french if TRUE}
}
\value{
an xtable

an xtable

an xtable

an xtable

an xtable

an xtable

an xtable

an xtable
}
\description{
Make a table of parameter estimates and priors

Make a table of reference points

Make a table for values such as biomass and recruitment

Make a table with both spawning biomass and depletion in it.
Based on [make.value.table()], but wider with both and extra headers

Make a table of the sensitivity parameter information as found in
the CSV file in the data directory

Make a table of the values of q, including quantiles

Make a table for the variance results data

Make a table of catchability parameters for sensitivities
}
