% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sens-params.R
\name{table_sens_params}
\alias{table_sens_params}
\title{Create a table of parameter estimates for multiple models}
\usage{
table_sens_params(
  tab,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\value{
an xtable
}
\description{
Make a table of the sensitivity parameter information as found in
the CSV file in the data directory
}
