% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R
\name{make.parameters.table}
\alias{make.parameters.table}
\alias{make.catchability.parameters.table}
\alias{param_est_mpd_table}
\alias{make.parameters.est.table}
\alias{make.ref.points.table}
\alias{make.value.table}
\alias{make.biomass.depletion.table}
\alias{make.sens.parameter.table}
\alias{make.sens.q.table}
\alias{make.variance.table}
\alias{make.catchability.parameters.table.q.sens}
\title{Make an xtable in the proper format for parameter estimates and priors}
\usage{
make.parameters.table(
  model,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.catchability.parameters.table(
  am1.lst,
  am2.lst = NULL,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

param_est_mpd_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = "5em",
  ...
)

make.parameters.est.table(
  model,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)

make.ref.points.table(
  model.am2,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  translate = FALSE
)

make.value.table(
  model,
  type,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)

make.biomass.depletion.table(
  model,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)

make.sens.parameter.table(
  tab,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.sens.q.table(
  models,
  model.names,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.variance.table(
  var,
  which.model = 1,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)

make.catchability.parameters.table.q.sens(
  qa.lst,
  qb.lst,
  qc.lst,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H"
)
}
\arguments{
\item{model}{an iscam model object}

\item{xcaption}{caption to appear in the calling document}

\item{xlabel}{the label used to reference the table in latex}

\item{font.size}{size of the font for the table}

\item{space.size}{size of the vertical spaces for the table}

\item{placement}{latex code for placement of the table in document}

\item{am1.lst}{A list of the AM1 iscam model objects}

\item{am2.lst}{A list of the AM2 iscam model objects}

\item{digits}{Numebr of digits to show}

\item{models}{A list of output models from [arrowtooth::model_setup()]}

\item{french}{If `TRUE` translate to French}

\item{...}{Arguments to pass to [csasdown::csas_table()]}

\item{translate}{Logical. Translate to french if TRUE}

\item{model.am2}{an iscam model object}

\item{type}{1=biomass, 2=recruitment, 3=F, 4=U, 5=depletion}

\item{syr}{start year for table}

\item{tabular.environment}{latex tabular environment value}

\item{tab}{the contents of the csv file as read in nby [readr::read_csv()]}

\item{model.names}{a vector of names of the models}

\item{var}{the variance parameter results data as loaded from
variance-parameter-results.csv}

\item{which.model}{which model to make the table for, 1 = AM1, 2 = AM2}

\item{qa.lst}{List of AM1 models for q sensitivities}

\item{qb.lst}{List of AM1 models for q sensitivities}

\item{qc.lst}{List of AM1 models for q sensitivities}
}
\value{
an [csasdown::csas_table()]

an xtable

an xtable

an xtable

an xtable

an xtable

an xtable

an xtable

an xtable
}
\description{
Make an xtable in the proper format for parameter estimates and priors

Make a tbale of the catchability parameters - Herring specific

Make a table of parameter estimate comparisons for MPD models

Make a table of parameter estimates and priors

Make a table of reference points

Make a table for values such as biomass and recruitment

Make a table with both spawning biomass and depletion in it.
Based on [make.value.table()], but wider with both and extra headers

Make a table of the sensitivity parameter information as found in
the CSV file in the data directory

Make a table of the values of q, including quantiles

Make a table for the variance results data

Make a table of catchability parameters for sensitivities
}
