% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscam-filenames.R
\docType{data}
\name{mcmc.sel.female.file}
\alias{mcmc.sel.female.file}
\title{iscam MCMC output female selectivity fits filename}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.sel.female.file
}
\description{
MCMC index female selectivity fits file
}
\keyword{datasets}
