% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfiscamutils.R
\docType{data}
\name{mcmc.sel.male.file}
\alias{mcmc.sel.male.file}
\title{The name of the ISCAM male selectivity file (csv)}
\format{
An object of class \code{character} of length 1.
}
\usage{
mcmc.sel.male.file
}
\description{
The name of the ISCAM male selectivity file (csv)
}
\keyword{datasets}
