% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{data}
\name{model.lst.class}
\alias{model.lst.class}
\title{iscam models list}
\format{An object of class \code{character} of length 1.}
\usage{
model.lst.class
}
\description{
Class definition
}
\keyword{datasets}
