% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{model_setup}
\alias{model_setup}
\title{Load models and set up lists and classes for the base model, bridge
model groups, and sensitivity model groups}
\usage{
model_setup(
  drs = NA,
  base_models_desc = NA,
  bridge_models_desc = NA,
  sens_models_desc = NA,
  retro_models_desc = NA,
  ...
)
}
\arguments{
\item{drs}{Output list from \code{\link[=set_dirs]{set_dirs()}}}

\item{base_models_desc}{A list of descriptions for the base models}

\item{bridge_models_desc}{A list of vectors of text strings to show in
the legends for bridge model plots, one name for each model, where the
list elements represent a group of models}

\item{sens_models_desc}{A list of vectors of text strings to show in
the legends for sensitivity model plots, one name for each model, where
the list elements represent a group of models}

\item{retro_models_desc}{A list of vectors of text strings to show in
the legends for retrospective model plots, one name for each model,
where the list elements represent a group of models}

\item{...}{Arguments to pass to \code{\link[=create_rds_file]{create_rds_file()}}}
}
\value{
A list of items, the base_model inside a single-element
list, the list of bridge model groups, sensitivity model groups, request
model groups, test model groups, and retrospective model groups. These
groups are lists of models which are to be compared with each other in
the document. This simplifies plotting and table functions
}
\description{
Load models and set up lists and classes for the base model, bridge
model groups, and sensitivity model groups
}
\details{
If any of the text lists are \code{NULL}, default description text will be
assigned and a warning given. If the object "models" is already in
existence, that will be used and only the names will be changed. This
allows for easy and fast switching between French and English.
}
