% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{model_setup}
\alias{model_setup}
\title{Load models and set up the directory names}
\usage{
model_setup(
  main_dirs = NULL,
  bridge_models_text = NULL,
  sens_models_text = NULL,
  ...
)
}
\arguments{
\item{main_dirs}{Output list from [set_dirs()]}

\item{bridge_models_text}{A list of vectors of text strings to show in the legends for bridge
model plots, one name for each model, where the list elements represent a group of models}

\item{sens_models_text}{A list of vectors of text strings to show in the legends for sensitivity
model plots, one name for each model, where the list elements represent a group of models}

\item{...}{Arguments to pass to [create_rds_file()]}
}
\value{
A list of three items, the base_model, the list of bridge model groups, and
the list of sensitivity model groups. The two lists are groups of models which are
to be compared with each other in the document. This simplifies plotting and table functions.
}
\description{
Load models and set up the directory names
}
\examples{
\dontrun{
library(gfiscamutils)
bridge_models_dirs <- c("01-base", "02-bridge-update-data")
sens_models_dirs <- list(c("01-base", "02-bridge-update-data"),
                         "01-base")
main_dirs <- set_dirs(base_model_dir = "base",
                      bridge_models_dirs = bridge_models_dirs,
                      sens_models_dirs = sens_models_dirs)
delete_files_ext(main_dirs$models, ext = "rds") # optional, shown for exposure
model_setup <- function(main_dirs,
                        overwrite_rds_files = TRUE)
}
}
