% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{move_legend_to_empty_facet}
\alias{move_legend_to_empty_facet}
\title{Moves a ggplot legend to an empty facet if one exists}
\usage{
move_legend_to_empty_facet(g)
}
\arguments{
\item{g}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Moves a ggplot legend to an empty facet if one exists
}
\details{
If an empty facet does not exist, an error will be thrown
If the input object \code{g} is not a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object nor a
\code{\link[ggplot2:ggplot_gtable]{ggplot2::ggplot_gtable()}} object, the original object is returned.
}
