% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{output.files}
\alias{output.files}
\title{iscam and ADMB output file patterns}
\format{
An object of class \code{character} of length 24.
}
\usage{
output.files
}
\description{
character vector
}
\keyword{datasets}
