% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R, R/tables-selectivity.R
\name{param_est_mpd_table}
\alias{param_est_mpd_table}
\alias{sel_param_est_mpd_table}
\title{Make a table of parameter estimate comparisons for MPD models.
Also show B0, Fmsy. Bmsy, and msy}
\usage{
param_est_mpd_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = "5em",
  ...
)

sel_param_est_mpd_table(
  models,
  digits = 3,
  french = FALSE,
  model_col_widths = NULL,
  ...
)
}
\arguments{
\item{models}{A list of iSCAM models as output from [load.iscam.files()]}

\item{digits}{Number of digits to show}

\item{french}{If `TRUE` translate to French}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to [csasdown::csas_table()]}
}
\value{
An [csasdown::csas_table()]

an [csasdown::csas_table()]
}
\description{
Make a table of parameter estimate comparisons for MPD models.
Also show B0, Fmsy. Bmsy, and msy

Make a table of selectivity parameter estimate comparisons for MPD models
}
