% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-parameters.R
\name{param_settings_table}
\alias{param_settings_table}
\title{Make a table showing number of parameters estimated and prior parameterizations}
\usage{
param_settings_table(model, french = FALSE, ...)
}
\arguments{
\item{model}{A single output model from [arrowtooth::model_setup()]}

\item{french}{If `TRUE` translate to French}

\item{...}{Arguments to pass to [csasdown::csas_table()]}
}
\description{
Make a table showing number of parameters estimated and prior parameterizations
}
