% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-fits-mcmc.R,
%   R/figures-age-residuals-mcmc.R
\name{plot_age_fits_mcmc}
\alias{plot_age_fits_mcmc}
\alias{plot_age_resids_mcmc}
\title{Plot the age fits for an MCMC model}
\usage{
plot_age_fits_mcmc(
  model,
  gear = 1,
  probs = c(0.025, 0.5, 0.975),
  comp_color = "black",
  comp_point_size = 0.5,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  text_title_size = 10,
  angle_x_labels = FALSE
)

plot_age_resids_mcmc(
  model,
  gear = 1,
  type = c("age", "year", "birth_year"),
  probs = c(0.025, 0.5, 0.975),
  comp_color = "black",
  comp_point_size = 0.5,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  ylim = c(-3, 3),
  text_title_size = 10,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{gear}{The number of the gear to plot}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age comp data points}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See `linetype` in [ggplot2]. Only used if `ci_type` is "line" or "both"}

\item{ci_color}{Color for the lines or ribbon for the credible intervals}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons. Only used if
`ci_type` is "ribbon" or "both"}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}

\item{type}{The type of residual plot to create. Options are "age",
"year", and "birth_year"}

\item{ylim}{For residual plots, set the y-axis limits with this two-element vector}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and 1 to anchor
the legend to. eg. c(1, 1) is the top right corner and c(0, 0) is the bottom left corner}

\item{title_text_size}{Add the model description as a title with this font size. The text
comes from the `model_desc` attribute of `model`. If this is `NULL`, don't show a title}
}
\value{
A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object
}
\description{
Plot the age fits for an MCMC model

Plot the age residuals for an MCMC model
}
\concept{Age plotting functions}
