% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-fits-mcmc.R
\name{plot_age_fits_mcmc}
\alias{plot_age_fits_mcmc}
\title{Plot MCMC age fits for an iSCAM model}
\usage{
plot_age_fits_mcmc(
  model,
  gear = 1,
  probs = c(0.025, 0.5, 0.975),
  yrs = NULL,
  comp_color = "black",
  comp_point_size = 0.5,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  text_title_size = 12,
  text_title_inc_mdl_nm = FALSE,
  angle_x_labels = FALSE,
  show_sample_size_f_only = TRUE,
  sample_size_x = 17,
  sample_size_y = 0.2
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The number of the gear to plot}

\item{probs}{A 3-element vector of probabilities that appear in the
output data frames. This is provided in case the data frames have more than
three different quantile levels}

\item{yrs}{A vector of years to include in the plot. If the maximum extends
past the range of years in the data, the maximum in the data will be used.
If \code{NULL}, all years will be included}

\item{comp_color}{Color for the age comp data lines and points}

\item{comp_point_size}{Size of the age comp data points}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See \code{linetype} in \link{ggplot2}. Only used if \code{ci_type} is
"line" or "both"}

\item{ci_color}{Color for the lines or ribbon for the credible intervals}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons.
Only used if \code{ci_type} is "ribbon" or "both"}

\item{text_title_inc_mdl_nm}{Logical. If \code{TRUE}, make the model name the
main title and the gear the subtitle. If \code{FALSE}, do not show the model
name, and make the gear the main title (if \code{text_title_size} is \code{NULL},
no title will be shown)}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}

\item{show_sample_size_f_only}{If \code{TRUE}, show the sample size in the female
panels only. If \code{FALSE}, show the sample size in both male and female panels.
Included for cases where the sample size is shared or repeated for both sexes.}

\item{sample_size_x}{The x value (age) to place the sample size label}

\item{sample_size_y}{The y value (0-1) to place the sample size label}

\item{type}{The type of residual plot to create. Options are "age", "year",
and "birth_year"}

\item{ylim}{For residual plots, set the y-axis limits with this two-element
vector}

\item{title_text_size}{Add the model description as a title with this font
size. The text comes from the \code{model_desc} attribute of \code{model}. If this is
\code{NULL}, don't show a title}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot MCMC age fits for an iSCAM model
}
\seealso{
Other Age plotting functions: 
\code{\link{plot_age_fits_mpd}()},
\code{\link{plot_age_resids_mcmc}()},
\code{\link{plot_age_resids_mpd}()}
}
\concept{Age plotting functions}
