% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-age-mpd.R
\name{plot_age_mpd}
\alias{plot_age_mpd}
\title{Plot the age fits for an MPD model}
\usage{
plot_age_mpd(
  model,
  gear = 1,
  type = c("fits", "resids"),
  resid_type = c("age", "year", "birth_year"),
  comp_color = "black",
  comp_point_size = 0.5,
  fir_line_color = "red",
  angle_x_labels = FALSE
)
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the age fits for an MPD model
}
