% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass-mcmc.R
\name{plot_biomass_mcmc}
\alias{plot_biomass_mcmc}
\title{Plot MCMC biomass for iSCAM models}
\usage{
plot_biomass_mcmc(
  models,
  rel = FALSE,
  show_bo = TRUE,
  probs = c(0.025, 0.5, 0.975),
  refpts_ribbon = TRUE,
  refpts_alpha = 0.2,
  bo_dodge = 0.1,
  x_space = 0.5,
  y_space = ifelse(rel, 0.05, 0.5),
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  angle_x_labels = FALSE,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{rel}{Logical. Make plot relative to initial estimate (B0), also known
as depletion}

\item{show_bo}{Logical. If \code{TRUE} and \code{rel == FALSE}, show the median
initial value on the plot (B0) as a point with the credible intervals as
bars}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{refpts_ribbon}{Logical. If \code{TRUE}, make the first model's reference
points lines (\code{show_bo_lines} and/or \code{show_bmsy_lines} must be \code{TRUE})
plotted an envelope of the credible interval, surrounding the median lines
for the reference points}

\item{refpts_alpha}{The transparency of the envelope shown for reference
points between 0 and 1. Only used when \code{refpts_ribbon} is \code{TRUE} and
\code{show_bo_lines} and/or \code{show_bmsy_lines} are \code{TRUE}}

\item{bo_dodge}{The amount to offset the initial value (B0 or R0) values
from each other so the values and uncertainty can be easily seen for
multiple models}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{y_space}{The amount of y-interval space to pad the top and bottom of
the plot with. To remove all padding, make this 0}

\item{show_bo_lines}{Show the B0 lines at values given by \code{bo_refpts} for
the first model in the \code{models} list}

\item{bo_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.2B0 and 0.4B0 by default}

\item{show_bmsy_lines}{Show the BMSY lines at values given by \code{bmsy_refpts}
for the first model in the \code{models} list}

\item{bmsy_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.4BMSY and 0.8BMSY by default}

\item{bo_refpt_colors}{A vector of two colors representing the LRP and USR
for B0. Used to display reference point lines if \code{show_bo_lines} is \code{TRUE}}

\item{bmsy_refpt_colors}{A vector of two colors representing the LRP and USR
for BMSY. Used to display reference point lines if \code{show_bmsy_lines} is \code{TRUE}}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}

\item{ylim}{A vector of two values, representing the minimum and maximum
values for the plot on the y-axis. If \code{NULL}, defaults will be zero for the
minimum and the ceiling of the maximum value (including CI)}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{...}{Arguments passed to \code{\link[=plot_ts_mcmc]{plot_ts_mcmc()}}}
}
\description{
Plot the MCMC spawning biomass trajectories for iscam models in either
absolute or relative form with or without reference point credible
intervals and medians
}
\details{
The reference point lines and credible intervals are output from the
first model in the list, as such they cannot be used to judge the other
model's status with respect to reference points. To see those, plot with
\code{\link[=plot_biomass_grid_mcmc]{plot_biomass_grid_mcmc()}} instead. This will show one panel per model,
each with its own reference points.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
