% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass-mpd.R
\name{plot_biomass_mpd}
\alias{plot_biomass_mpd}
\title{Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.}
\usage{
plot_biomass_mpd(
  models,
  rel = FALSE,
  show_bo = TRUE,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  alpha = 0.2,
  palette = "Paired",
  base_color = "black",
  bo_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ind_letter = NULL,
  leg_loc = NULL,
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{legend_title}{Title for legend}

\item{xlim}{The x limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{alpha}{The transparency with values from 0 to 1 of the ribbon shading
when \code{first_model_ribbon} is \code{TRUE}}

\item{palette}{A palette value that is in \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}}

\item{base_color}{A color to prepend to the brewer colors which are set by
\code{palette}. This is called \code{base_color} because it is likely to be a base
model}

\item{bo_dodge}{The amount to offset the initial value (B0 or R0) values
from each other so the values and uncertainty can be easily seen for
multiple models}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
