% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catches.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catch from a data frames as extracted from iscam data (dat) files}
\usage{
plot_catch(model, gear = 1, area = 1, group = 1, sex = 0,
  type = 1, thm = theme_bw(), translate = FALSE)
}
\arguments{
\item{model}{a gfiscamutils model object}

\item{gear}{gear number as it appears in iscam data file}

\item{area}{area number as it appears in iscam data file}

\item{group}{group number as it appears in iscam data file}

\item{sex}{sex number as it appears in iscam data file}

\item{type}{type number as it appears in iscam data file}

\item{thm}{[ggplot::theme()] function to use}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
A ggplot object
}
\description{
Plot catch from a data frames as extracted from iscam data (dat) files
}
