% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-distributions.R
\name{plot_fun}
\alias{plot_fun}
\title{Plot a distribution function. Mostly for use with making prior plots
for MCMC parameters. See \code{\link[=plot_priors_posts_mcmc]{plot_priors_posts_mcmc()}}}
\usage{
plot_fun(
  fun = dnorm,
  nsamp = 100,
  param_lst = list(mean = 0, sd = 0.5),
  xlim = c(-2, 2),
  fill = "steelblue",
  title = "",
  mode_line_col = "red",
  mode_line_lwd = 0.5,
  mode_line_lty = 1,
  mean_line_col = "black",
  mean_line_lwd = 0.5,
  mean_line_lty = 1,
  sd_lines_col = "black",
  sd_lines_lwd = 0.5,
  sd_lines_lty = 2,
  alpha = 0.75,
  show_mode_line = TRUE,
  show_mean_line = TRUE,
  show_sd_lines = TRUE
)
}
\arguments{
\item{fun}{The function to plot}

\item{nsamp}{Number of points across x-axis}

\item{param_lst}{A list of parameters required by the function provided (\code{fun})
They can be named but do not have to be}

\item{xlim}{x-axis limits (2-element vector)}

\item{fill}{Fill color for under the curve}

\item{title}{Plot title}

\item{mode_line_col}{Mode line color}

\item{mode_line_lwd}{Mode line width}

\item{mode_line_lty}{Mode line type}

\item{mean_line_col}{Mean line color}

\item{mean_line_lwd}{Mean line width}

\item{mean_line_lty}{Mean line type}

\item{sd_lines_col}{Standard deviation lines color}

\item{sd_lines_lwd}{Standard deviation lines width}

\item{sd_lines_lty}{Standard deviation lines type}

\item{alpha}{Opacity of fill from 0-1}

\item{show_mode_line}{Logical. If \code{TRUE}, show a vertical line at the mode
of the distribution}

\item{show_mean_line}{Logical. If \code{TRUE}, show a vertical line at
the mean}

\item{show_sd_lines}{Logical. If \code{TRUE}, show vertical lines one standard
deviation away from the mean}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a distribution function. Mostly for use with making prior plots
for MCMC parameters. See \code{\link[=plot_priors_posts_mcmc]{plot_priors_posts_mcmc()}}
}
\details{
The plot will contain a filled curve representing the prior distribution,
and a solid red (\code{mode_line_col}) vertical line representing the mode of the distribution and a
solid black (\code{mean_line_col}) and two dotted black (\code{sd_lines_col}) vertical lines
representing the mean and one standard deviation from the mean
}
