% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-indices.R
\name{plot_index_fits_mcmc}
\alias{plot_index_fits_mcmc}
\title{Plot MCMC index fits}
\usage{
plot_index_fits_mcmc(
  models,
  model_names = factor(names(models), levels = names(models)),
  surv_index,
  start_year = 1996,
  end_year = 2021,
  legend_title = "Bridge model",
  palette = "Paired",
  line_width = 0.5,
  point_size = 1
)
}
\description{
Plot MCMC index fits
}
