% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-index-mpd.R
\name{plot_index_mpd}
\alias{plot_index_mpd}
\title{Plot MPD index fits for a single or group of models}
\usage{
plot_index_mpd(
  models,
  type = c("fits", "resids"),
  surv_index,
  gear = NULL,
  start_year = 1995,
  end_year = 2021,
  append_base_txt = NULL,
  legend_title = "Models",
  palette = "Paired",
  base_color = "black",
  dodge = 0.3,
  index_line_width = 0.5,
  index_point_size = 2,
  index_color = "chocolate3",
  fit_line_width = 0.5,
  fit_point_size = 2,
  leg_loc = c(1, 1),
  angle_x_labels = FALSE
)
}
\description{
Plot MPD index fits for a single or group of models
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()}
}
\concept{Time series plotting functions}
