% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-index-mcmc.R, R/figures-index-mpd.R
\name{plot_index_mcmc}
\alias{plot_index_mcmc}
\alias{plot_index_mpd}
\title{Plot MCMC index fits for a single or group of models}
\usage{
plot_index_mcmc(
  models,
  model_names = NULL,
  type = c("fits", "resids"),
  surv_index,
  start_year = 1995,
  end_year = 2021,
  legend_title = "Models",
  palette = "Paired",
  dodge = 0.3,
  index_line_width = 0.5,
  index_point_size = 2,
  fit_line_width = 0.5,
  fit_point_size = 2,
  errbar_width = 0.5,
  angle_x_labels = FALSE
)

plot_index_mpd(
  models,
  model_names = NULL,
  type = c("fits", "resids"),
  surv_index,
  start_year = 1995,
  end_year = 2021,
  legend_title = "Models",
  palette = "Paired",
  dodge = 0.3,
  index_line_width = 0.5,
  index_point_size = 2,
  fit_line_width = 0.5,
  fit_point_size = 2,
  angle_x_labels = FALSE
)
}
\arguments{
\item{models}{A list of iscam model objects (class [mdl_lst_cls])}

\item{model_names}{Names to use for the models in the plots. The names of
the list items in `models` will be used if they are present and this will
be ignored. If the list item names are not defined, temporary names will be used}

\item{surv_index}{The `survey_index` data frame which is the `dat` object in the output
from the [read_data_file()] function}

\item{start_year}{Year to start plot}

\item{end_year}{Year to end plot}

\item{legend_title}{Title text for the legend}

\item{palette}{A palette value that is in [RColorBrewer::brewer.pal.info]}

\item{dodge}{A small value added to each year for each model. This is added cumulatively,
so each model fit appears more to the right than the previous one}

\item{index_line_width}{The index data error bar and connecting line width}

\item{index_point_size}{The index data point size}

\item{fit_line_width}{The model fit error bar and connecting line width}

\item{fit_point_size}{The model fit point size}

\item{errbar_width}{The width of the top and bottom crossbar of the errorbars}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot MCMC index fits for a single or group of models

Plot MPD index fits for a single or group of models
}
\concept{index plotting functions}
