% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-priors-posts-mcmc.R
\name{plot_priors_posts_mcmc}
\alias{plot_priors_posts_mcmc}
\title{Plot MCMC priors and posteriors for iSCAM models}
\usage{
plot_priors_posts_mcmc(
  model,
  param_rm = c("sel", "bo", "tau", "sigma"),
  priors_only = FALSE,
  text_title_size = 12,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If \code{NULL}, nothing is removed}

\item{priors_only}{Logical. If \code{TRUE}, plot the priors only, not the
posteriors}

\item{text_title_size}{Add the model description as a title with this font size. The text
comes from the \code{model_desc} attribute of \code{model}. If this is \code{NULL}, don't show a title}

\item{...}{Additional arguments passed to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}}
}
\description{
Priors and posteriors comparison plots for parameters in an MCMC run
}
\details{
Plots the priors overlaid on the posteriors for the given iscam model.
The values in the control file (\code{model$ctl$params}) for each
prior are:
\enumerate{
\item ival  = initial value
\item lb    = lower bound
\item ub    = upper bound
\item phz   = ADMB phase
\item prior = prior distribution funnction
0 = Uniform
1 = normal    (p1 = mu, p2 = sig)
2 = lognormal (p1 = log(mu), p2 = sig)
3 = beta      (p1 = alpha, p2 = beta)
4 = gamma     (p1 = alpha, p2 = beta)
\item p1 (defined by 5 above)
\item p2 (defined by 5 above)
}

See \code{\link[=plot_distribution]{plot_distribution()}} for prior plot details including an documentation
on the vertical lines
}
\seealso{
Other MCMC diagnostics plots: 
\code{\link{plot_autocorr_mcmc}()},
\code{\link{plot_pairs_mcmc}()},
\code{\link{plot_traces_mcmc}()}
}
\concept{MCMC diagnostics plots}
