% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-q-mcmc.R
\name{plot_q_mcmc}
\alias{plot_q_mcmc}
\title{Plot MCMC catchabilities for iSCAM models}
\usage{
plot_q_mcmc(
  models,
  gear = 1,
  probs = c(0.025, 0.5, 0.975),
  base_color = "black",
  palette = "Paired",
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_color = "red",
  ci_alpha = 0.3,
  text_title_size = 12,
  angle_x_labels = FALSE
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{gear}{The number of the gear to plot}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{base_color}{A color to prepend to the brewer colors which are set by
\code{palette}. This is called \code{base_color} because it is likely to be a base
model}

\item{palette}{A palette value that is in \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}
}
\description{
Plot catchability plots for MCMC output with credible intervals
}
\details{
This function can plot either multiple gears for a single model on one plot,
or a single gear from multiple models on a single plot. Errors will be
thrown if trying to plot multiple gears and multiple models together. To
do that, call \code{\link[=plot_q_grid_mcmc]{plot_q_grid_mcmc()}} instead.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
