% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ref-points.R
\name{plot_ref_points}
\alias{plot_ref_points}
\title{Plot reference point boxplots}
\usage{
plot_ref_points(
  model,
  ref_pts = c("sbo", "bmsy", "msy1", "msy2", "umsy1", "umsy2"),
  ref_desc = c(expression(italic(SB)[0]), expression(italic(B)[MSY]),
    expression(italic(MSY)[1]), expression(italic(MSY)[2]), expression(italic(UMSY)[1]),
    expression(italic(UMSY)[2])),
  color = "black",
  fill = "steelblue"
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}
}
\value{
A \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} box plots
}
\description{
Plot reference point boxplots
}
