% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selex-gear-mcmc.R
\name{plot_selex_gear_mcmc}
\alias{plot_selex_gear_mcmc}
\title{Plot MCMC selectivities by single gear for iSCAM models}
\usage{
plot_selex_gear_mcmc(
  model,
  gear = 1,
  probs = c(0.025, 0.5, 0.975),
  show_maturity = FALSE,
  ci_type = c("both", "line", "ribbon"),
  ci_linetype = c("dotted", "solid", "dashed", "dotdash", "longdash", "twodash"),
  ci_alpha = 0.3
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{gear}{The gear number in the model}

\item{probs}{A 3-element vector of probabilities that appear in the output data frames
This is provided in case the data frames have more than three different quantile levels}

\item{show_maturity}{If \code{TRUE}, overlay the maturity ogive on the selectivity plots}

\item{ci_type}{One of "line", "ribbon", "both" to show the credible interval}

\item{ci_linetype}{See \code{linetype} in \link{ggplot2}. Only used if \code{ci_type} is "line" or "both"}

\item{ci_alpha}{Opacity between 0 and 1 for the credible intervals ribbons. Only used if
\code{ci_type} is "ribbon" or "both"}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the (possibly time-varying) selectivity for a particular gear in
an iscam model
}
\seealso{
Other Selectivity plotting functions: 
\code{\link{plot_selex_mcmc}()},
\code{\link{plot_selex_mpd}()},
\code{\link{plot_tv_selex_mpd}()}
}
\concept{Selectivity plotting functions}
