% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-selectivity.R
\name{plot_selex_mpd}
\alias{plot_selex_mpd}
\title{Plot the selectivity for all gears in the iscam model}
\usage{
plot_selex_mpd(
  model,
  show_maturity = FALSE,
  last_year_only = TRUE,
  title = NULL
)
}
\arguments{
\item{model}{An iscam model object as output by [arrowtooth::model_setup()]}

\item{show_maturity}{If `TRUE`, overlay the maturity ogive on the selectivity plots}

\item{last_year_only}{If `TRUE`, only show the last year. This would be used for non-time-varying selectivities}

\item{title}{The plot title. NULL means no title}
}
\value{
A [ggplot2::ggplot()] object
}
\description{
Plot the selectivity for all gears in the iscam model
}
