% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sr-mcmc.R
\name{plot_sr_mcmc}
\alias{plot_sr_mcmc}
\title{Plot the stock-recruitment relationship for an MCMC ISCAM model}
\usage{
plot_sr_mcmc(model, probs = c(0.025, 0.5, 0.975), errorbar_thickness = 0.1)
}
\arguments{
\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}
}
\description{
Plot the stock-recruitment relationship for an MCMC ISCAM model
}
