% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-traces-mcmc.R, R/figures-autocor-mcmc.R,
%   R/figures-pairs.R
\name{plot_traces_mcmc}
\alias{plot_traces_mcmc}
\alias{plot_autocor}
\alias{plot_pairs_mcmc}
\title{Trace plots for parameters in an MCMC run}
\usage{
plot_traces_mcmc(
  model,
  plot_sel = FALSE,
  param_rm = NULL,
  list_param_names = FALSE,
  ...
)

plot_autocor(
  model,
  plot_sel = FALSE,
  param_rm = NULL,
  lag = 100,
  list_param_names = FALSE,
  ...
)

plot_pairs_mcmc(
  model,
  plot_sel = FALSE,
  param_rm = c("rho", "vartheta"),
  list_param_names = FALSE
)
}
\arguments{
\item{model}{An iscam model object (class [mdl_cls])}

\item{plot_sel}{Logical. If `TRUE`, plot only the selectivity parameters.
If `FALSE`, plot only non-selectivity parameters}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If `NULL`, nothing is removed}

\item{list_param_names}{Logical. If `TRUE`, list the parameter names that will be plot
by this function, but do not actually plot it. This is to help you to use the
`param_rm` argument, so you know what the names to be removed can be}

\item{...}{Additional arguments passed to [cowplot::plot_grid()]}
}
\value{
A [ggplot2::gglot()] object

A [ggplot2::ggplot()] object made from the [forecast::ggAcf()] function

A [ggplot2::ggplot()] object made from the [GGally::ggpairs()] function
}
\description{
Trace plots for parameters in an MCMC run

Autocorrelation plots for parameters in an MCMC run

Pairs plots for parameters in an MCMC run
}
\details{
Show correlations between parameters as scatterplots in the lower
triangular area, density of each parameter in the diagonals, and the correlation
value with text size scaled to be larger for higher correlations in the upper
triangular area
}
