% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-traces-mcmc.R
\name{plot_traces_mcmc}
\alias{plot_traces_mcmc}
\title{Plot MCMC traces for iSCAM models}
\usage{
plot_traces_mcmc(
  model,
  plot_sel = FALSE,
  param_rm = c("rho", "vartheta"),
  list_param_names = FALSE,
  text_title_size = 12,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{plot_sel}{Logical. If \code{TRUE}, plot only the selectivity parameters.
If \code{FALSE}, plot only non-selectivity parameters. If \code{NULL}, plot all parameters
except those in \code{param_rm}}

\item{param_rm}{A vector of parameter names to remove. If any are not found
in the output, a warning will be issued. If \code{NULL}, nothing is removed}

\item{list_param_names}{Logical. If \code{TRUE}, list the parameter names that will be plot
by this function, but do not actually plot it. This is to help you to use the
\code{param_rm} argument, so you know what the names to be removed can be}

\item{text_title_size}{Add the model description as a title with this font size. The text
comes from the \code{model_desc} attribute of \code{model}. If this is \code{NULL}, don't show a title}

\item{...}{Additional arguments passed to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Trace plots for parameters in an MCMC run
}
\seealso{
Other MCMC diagnostics plots: 
\code{\link{plot_autocorr_mcmc}()},
\code{\link{plot_pairs_mcmc}()},
\code{\link{plot_priors_posts_mcmc}()}
}
\concept{MCMC diagnostics plots}
