% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass.R
\name{plot_ts_mcmc}
\alias{plot_ts_mcmc}
\title{Plot the biomass trajectories for iscam models}
\usage{
plot_ts_mcmc(
  models,
  model_names = NULL,
  type = "sbt",
  rel = FALSE,
  legend_title = "Bridge model",
  palette = "Paired",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  lineribbon = FALSE,
  alpha = 0.5,
  offset = 0.1,
  bo_dodge = 0.05,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bmsy_line = FALSE,
  show_bo_line = TRUE,
  ind_letter = NULL,
  leg = NULL,
  probs = c(0.025, 0.5, 0.975),
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects}

\item{ylim}{The y limits for the plot}

\item{offset}{The amount on the x-axis to offset each point and line for
multiple models}

\item{leg}{Position of the legend. NULL means no legend is shown}

\item{...}{Other graphical arguments}

\item{model.names}{A vector of names to show on the legend}

\item{opacity}{How opaque the credibility envelopes are}

\item{append.base.txt}{Text to append to the name of the first model}

\item{show.bmsy.line}{Show the reference lines 0.4 and 0.8bmsy}

\item{show.bo.line}{Show the reference lines 0.2 and 0.4bo}

\item{ind.letter}{A letter to show on the plot (for panel plots)}
}
\value{
Nothing
}
\description{
Plot the biomass with credibility intervals for the mcmc
cases of the models
}
