% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ts-mcmc.R
\name{plot_ts_mcmc}
\alias{plot_ts_mcmc}
\title{Plot MCMC time series plots for iSCAM models}
\usage{
plot_ts_mcmc(
  models,
  quant_df = "sbt_quants",
  facet_wrap_var = c("none", "sex", "gear"),
  palette = iscam_palette,
  all_one_color = NULL,
  base_color = "black",
  legend_title = ifelse(fr(), "Modèles", "Models"),
  x_label = ifelse(fr(), "Année", "Year"),
  y_label = ifelse(fr(), "Biomasse reproductrice (milliers de t)",
    "Spawning Biomass (thousand t)"),
  append_base_txt = NULL,
  x_space = 0.5,
  y_space = 0,
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  first_model_ribbon = TRUE,
  alpha = 0.2,
  leg_loc = c(1, 1),
  probs = c(0.025, 0.5, 0.975),
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{quant_df}{A name of a value which can be found in the
\code{model$mcmccalcs} list for each iscam model in the \code{models} list}

\item{facet_wrap_var}{A column in the \code{quant_df} data frame which will be
used to split the output into separate panel plots. If "none", no faceting
will be done}

\item{palette}{A palette value that is in \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info}}

\item{all_one_color}{If not \code{NULL}, overrides \code{palette} and the color that
this is defined as will be used for all lines on the plot (all models)}

\item{base_color}{A color to prepend to the brewer colors which are set by
\code{palette}. This is called \code{base_color} because it is likely to be a base
model}

\item{legend_title}{Title for legend}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{y_space}{The amount of y-interval space to pad the top and bottom of
the plot with. To remove all padding, make this 0}

\item{xlim}{The x limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{first_model_ribbon}{Logical. If \code{TRUE}, give the first model a shaded
credible interval instead of dotted lines}

\item{alpha}{The transparency with values from 0 to 1 of the ribbon shading
when \code{first_model_ribbon} is \code{TRUE}}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}

\item{bo_dodge}{The amount to offset the initial value (B0 or R0) values
from each other so the values and uncertainty can be easily seen for
multiple models}

\item{title_text_size}{Add the model description as a title with this font
size. The text comes from the \code{model_desc} attribute of \code{model}. If this is
\code{NULL}, don't show a title}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot time series plots for MCMC output with credible intervals for single or
multiple models
}
\details{
Plot a single model or multiple models overlaid for comparison.
If a single model, it must have a class attribute of \code{mdl_cls}.
A list of multiple models must have a class attribute of \code{mdl_lst_cls} and
each model in the list must have a class attribute of \code{mdl_cls}.

The model description text will appear in a legend if \code{models} length
is greater than 1, and as a title if there is only one model. The
description text is found in the \code{model_desc} attribute of a
\code{mdl_cls} class ("iscam_model") object. The attribute is set in the
\code{\link[=model_setup]{model_setup()}} function, from either the \code{bridge_model_text} or
\code{sens_model_text} arguments.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_biomass_proj_mcmc}()},
\code{\link{plot_catch_fit_mcmc}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_vuln_mcmc}()}
}
\concept{Time series plotting functions}
