% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-biomass-grid-mcmc.R,
%   R/figures-biomass-mcmc.R, R/figures-biomass-mpd.R, R/figures-f-mcmc.R,
%   R/figures-recruitment-grid-mcmc.R, R/figures-recruitment-mcmc.R,
%   R/figures-recruitment-mpd.R, R/figures-time-series-mcmc.R,
%   R/figures-vuln-biomass-mcmc.R
\name{plot_biomass_grid_mcmc}
\alias{plot_biomass_grid_mcmc}
\alias{plot_biomass_mcmc}
\alias{plot_biomass_mpd}
\alias{plot_f_mcmc}
\alias{plot_recr_grid_mcmc}
\alias{plot_recr_mcmc}
\alias{plot_recr_mpd}
\alias{plot_ts_mcmc}
\alias{plot_vuln_mcmc}
\title{Plot a grid of biomass plots, each with one model only so that the reference
point lines and credible intervals can be compared correctly.}
\usage{
plot_biomass_grid_mcmc(models, ...)

plot_biomass_mcmc(
  models,
  rel = FALSE,
  show_bo = TRUE,
  probs = c(0.025, 0.5, 0.975),
  refpts_ribbon = TRUE,
  refpts_alpha = 0.2,
  bo_dodge = 0.1,
  x_space = 0.5,
  y_space = ifelse(rel, 0.05, 0.5),
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ...
)

plot_biomass_mpd(
  models,
  rel = FALSE,
  show_bo = TRUE,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  alpha = 0.2,
  palette = "Paired",
  base_color = "black",
  bo_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  show_bo_lines = FALSE,
  bo_refpts = c(0.2, 0.4),
  show_bmsy_lines = FALSE,
  bmsy_refpts = c(0.4, 0.8),
  bo_refpt_colors = c("red", "green"),
  bmsy_refpt_colors = c("salmon", "darkgreen"),
  ind_letter = NULL,
  leg_loc = NULL,
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)

plot_f_mcmc(..., type = c("i", "e"))

plot_recr_grid_mcmc(models, ...)

plot_recr_mcmc(
  models,
  show_ro = TRUE,
  ro_color = base_color,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  ro_ribbon = TRUE,
  ro_alpha = 0.3,
  palette = "Paired",
  base_color = "black",
  r_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)

plot_recr_mpd(
  models,
  model_names = NULL,
  show_ro = TRUE,
  ro_color = base_color,
  legend_title = "Models",
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  alpha = 0.2,
  palette = "Paired",
  base_color = "#000000",
  r_dodge = 0.1,
  x_space = 0.5,
  append_base_txt = NULL,
  ind_letter = NULL,
  leg_loc = NULL,
  probs = c(0.025, 0.5, 0.975),
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)

plot_ts_mcmc(
  models,
  quant_df = "sbt",
  facet_wrap_var = c("none", "sex", "gear"),
  palette = "Paired",
  base_color = "black",
  legend_title = "Models",
  x_label = "Year",
  y_label = "Spawning Biomass ('000 t)",
  append_base_txt = NULL,
  x_space = 0.5,
  y_space = 0,
  xlim = NULL,
  ylim = NULL,
  line_width = 1,
  point_size = 2,
  first_model_ribbon = TRUE,
  alpha = 0.2,
  leg_loc = c(1, 1),
  probs = c(0.025, 0.5, 0.975),
  text_title_size = 12,
  angle_x_labels = FALSE,
  ...
)

plot_vuln_mcmc(...)
}
\arguments{
\item{models}{A list of iscam model objects (class [mdl_lst_cls])}

\item{...}{Arguments passed to [plot_ts_mcmc()]}

\item{rel}{Logical. Make plot relative to initial estimate (B0), also known
as depletion}

\item{show_bo}{Logical. If `TRUE` and `rel == FALSE`, show the median
initial value on the plot (B0) as a point with the credible intervals as
bars}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{refpts_ribbon}{Logical. If `TRUE`, make the first model's reference
points lines (`show_bo_lines` and/or `show_bmsy_lines` must be `TRUE`)
plotted an envelope of the credible interval, surrounding the median lines
for the reference points}

\item{refpts_alpha}{The transparency of the envelope shown for reference
points between 0 and 1. Only used when `refpts_ribbon` is `TRUE` and
`show_bo_lines` and/or `show_bmsy_lines` are `TRUE`}

\item{bo_dodge}{The amount to offset the initial value (B0 or R0) values
from each other so the values and uncertainty can be easily seen for
multiple models}

\item{x_space}{The amount of x-interval space to pad the left and right of
the plot with. To remove all padding, make this 0}

\item{y_space}{The amount of y-interval space to pad the top and bottom of
the plot with. To remove all padding, make this 0}

\item{show_bo_lines}{Show the B0 lines at values given by `bo_refpts` for
the first model in the `models` list}

\item{bo_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.2B0 and 0.4B0 by default}

\item{show_bmsy_lines}{Show the BMSY lines at values given by `bmsy_refpts`
for the first model in the `models` list}

\item{bmsy_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.4BMSY and 0.8BMSY by default}

\item{bo_refpt_colors}{A vector of two colors representing the LRP and USR
for B0. Used to display reference point lines if `show_bo_lines` is `TRUE`}

\item{bmsy_refpt_colors}{A vector of two colors representing the LRP and USR
for BMSY. Used to display reference point lines if `show_bmsy_lines` is `TRUE`}

\item{legend_title}{Title for legend}

\item{xlim}{The x limits for the plot. If `NULL`, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If `NULL`, the limits of the data
will be used}

\item{line_width}{Width of all median lines on the plot}

\item{point_size}{Point size for all median points on the plot}

\item{alpha}{The transparency with values from 0 to 1 of the ribbon shading
when `first_model_ribbon` is `TRUE`}

\item{palette}{A palette value that is in [RColorBrewer::brewer.pal.info]}

\item{base_color}{A color to prepend to the brewer colors which are set by
`palette`. This is called `base_color` because it is likely to be a base
model}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}

\item{angle_x_labels}{If `TRUE` put 45 degree angle on x-axis tick labels}

\item{type}{'i' means instantaneous, 'e' means exploitation rate (Ut)}

\item{show_ro}{Show the initial recruitment, R0 median line and credible interval}

\item{ro_ribbon}{See `refpts_ribbon` in [plot_biomass_mcmc()]}

\item{ro_alpha}{See `refpts_alpha` in [plot_biomass_mcmc()]}

\item{r_dodge}{See `bo_dodge` in [plot_biomass_mcmc()]}

\item{quant_df}{A value which can be found in the `model$mcmccalcs` list
for each iscam model in the `models` list}

\item{facet_wrap_var}{A column in the `quant_df` data frame which will be
used to split the output into separate panel plots. If "none", no faceting
will be done}

\item{first_model_ribbon}{Logical. If `TRUE`, give the first model a shaded
credible interval instead of dotted lines}

\item{title_text_size}{Add the model description as a title with this font
size. The text comes from the `model_desc` attribute of `model`. If this is
`NULL`, don't show a title}
}
\value{
A [cowplot::plot_grid()] object

A [ggplot2::ggplot()] object

A [cowplot::plot_grid()] object

A [ggplot2::ggplot()] object

A [ggplot2::ggplot()] object
}
\description{
Plot a grid of biomass plots, each with one model only so that the reference
point lines and credible intervals can be compared correctly.

Plot the MCMC spawning biomass trajectories for iscam models in either
absolute or relative form with or without reference point credible
intervals and medians

Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.

Plot fishing mortality time series plots for MCMC output with credible
intervals

Plot a grid of recruitment plots, each with one model only so that
the R0 line and credible intervals can be viewed

Plot the MCMC recruitment time series trajectories for iscam models.

Plot the MCMC time series trajectories for iscam models, including spawning biomass
and recruitment for both absolute and relative cases.

Plot time series plots for MCMC output with credible intervals for single or
multiple models

Plot vulnerable biomass time series plots for MCMC output with credible
intervals
}
\details{
The reference point lines and credible intervals are output from the
first model in the list, as such they cannot be used to judge the other
model's status with respect to reference points. To see those, plot with
[plot_biomass_grid_mcmc()] instead. This will show one panel per model,
each with its own reference points.

Plot a single model or multiple models overlaid for comparison.
If a single model, it must have a class attribute of `mdl_cls`.
A list of multiple models must have a class attribute of `mdl_lst_cls` and
each model in the list must have a class attribute of `mdl_cls`.

The model description text will appear in a legend if `models` length
is greater than 1, and as a title if there is only one model. The
description text is found in the `model_desc` attribute of a
`mdl_cls` class ("iscam_model") object. The attribute is set in the
[model_setup()] function, from either the `bridge_model_text` or
`sens_model_text` arguments.
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}

Other Time series plotting functions: 
\code{\link{plot_index_mcmc}()}
}
\concept{Time series plotting functions}
