% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vuln-mcmc.R
\name{plot_vuln_mcmc}
\alias{plot_vuln_mcmc}
\title{Plot MCMC vulnerable biomass for iSCAM models}
\usage{
plot_vuln_mcmc(model, xlim = NULL, ylim = NULL, angle_x_labels = FALSE)
}
\arguments{
\item{xlim}{The x limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels}

\item{inc_sbt}{Logical. If \code{TRUE}, put all gear's vulnerable biomass
on one plot along with the spawning biomass (sbt). If \code{FALSE}, make
a panel plot with one vulnerable biomass plot per panel, with no
spawning biomass}
}
\description{
Plot vulnerable biomass time series plots for MCMC output with credible
intervals
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()}
}
\concept{Time series plotting functions}
