% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vuln-mcmc.R
\name{plot_vuln_mcmc}
\alias{plot_vuln_mcmc}
\title{Plot MCMC vulnerable biomass for iSCAM models}
\usage{
plot_vuln_mcmc(
  model,
  append_base_txt = tr("Spawning biomass"),
  xlim = NULL,
  ylim = NULL,
  leg_loc = NULL,
  angle_x_labels = FALSE,
  units = c("kt", "1000 t")
)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{append_base_txt}{A vector of strings to append to the model names for
display on the plot legend or title}

\item{xlim}{The x limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{ylim}{The y limits for the plot. If \code{NULL}, the limits of the data
will be used}

\item{leg_loc}{A two-element vector describing the X-Y values between 0 and
1 to anchor the legend to. eg. c(1, 1) is the top right corner and c(0, 0)
is the bottom left corner}

\item{angle_x_labels}{If \code{TRUE} put 45 degree angle on x-axis tick labels
minimum and the ceiling of the maximum value (including CI)}

\item{units}{One of "1000 t" or "kt". The text that will appear in the
y-axis label. The "1000 t" text will be changed to "1,000 t" or "1 000 t"
for English or French respectively}
}
\description{
Plot vulnerable biomass time series plots for MCMC output with credible
intervals
}
\seealso{
Other Time series plotting functions: 
\code{\link{plot_biomass_grid_mcmc}()},
\code{\link{plot_biomass_mcmc}()},
\code{\link{plot_biomass_mpd}()},
\code{\link{plot_biomass_proj_mcmc}()},
\code{\link{plot_catch_fit_mcmc}()},
\code{\link{plot_f_mcmc}()},
\code{\link{plot_index_mcmc}()},
\code{\link{plot_index_mpd}()},
\code{\link{plot_q_mcmc}()},
\code{\link{plot_rdevs_mcmc}()},
\code{\link{plot_recr_grid_mcmc}()},
\code{\link{plot_recr_mcmc}()},
\code{\link{plot_recr_mpd}()},
\code{\link{plot_ts_mcmc}()}
}
\concept{Time series plotting functions}
