% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{rdata.file}
\alias{rdata.file}
\title{iscam model RData file name}
\format{An object of class \code{character} of length 1.}
\usage{
rdata.file
}
\description{
character
}
\keyword{datasets}
