% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{read_data_file}
\alias{read_data_file}
\title{Read in the iscam data file}
\usage{
read_data_file(file = NULL, verbose = FALSE)
}
\arguments{
\item{file}{Filename (full path)}

\item{verbose}{Say more}
}
\value{
A list representing the contents of the iscam data file
}
\description{
Read in the iscam data file
}
