% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{retro.dir}
\alias{retro.dir}
\title{Retrospective directory name}
\format{An object of class \code{character} of length 1.}
\usage{
retro.dir
}
\description{
character
}
\keyword{datasets}
