% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{set_dirs}
\alias{set_dirs}
\title{Set main directories for the project. Check existence of all directories and report.
Sensitivity groups will have the base model prepended.}
\usage{
set_dirs(
  nongit_dir = file.path(dirname(here()), paste0(basename(here()), "-nongit")),
  models_dir = file.path(nongit_dir, "models"),
  base_model_dir = "base",
  bridge_models_dir = "001-bridge-models",
  bridge_models_dirs = NULL,
  sens_models_dir = "002-sens-models",
  sens_models_dirs = NULL
)
}
\arguments{
\item{nongit_dir}{The full path containing non-version-controlled things
such as data, model runs from iSCAM, and reference materials. Default is "reponame-nongit"}

\item{models_dir}{The full path in which the iSCAM model directories are located}

\item{base_model_dir}{Name of the base model directory which is a subdirectory of \code{models}}

\item{bridge_models_dir}{Name of the subdirectory of \code{models_dir} that
contains the bridge model directories}

\item{bridge_models_dirs}{A vector of subdirectory names in \code{models_dir/bridge_models_dir}
that each contain an individual iSCAM bridge model}

\item{sens_models_dir}{Name of the subdirectory of \code{models_dir} that
contains the sensitivity model directories}

\item{sens_models_dirs}{A vector of subdirectory names in \code{models_dir/sens_models_dir}
that each contain an individual iSCAM sensitivity model}
}
\value{
A list of seven items, the first two are the same as the input arguments with the same name.
The 3rd is the full path of the base model.
The 4th is the full path of the bridge models directory.
The 5th contains a vector of the full paths to the bridge models which are inside the 4th directory.
The 6th is the full path of the sensitivity models directory.
The 7th contains a vector of the full paths to the sensitivity models which are inside the 6th directory.
}
\description{
Set main directories for the project. Check existence of all directories and report.
Sensitivity groups will have the base model prepended.
}
