% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-gear-language.R
\name{set_gear_language}
\alias{set_gear_language}
\title{Translate the gear names and abbreviations in the models list for all
models to either French of English}
\usage{
set_gear_language(models)
}
\arguments{
\item{models}{A list of the models}
}
\value{
A modified \code{models} list of class \code{mdl_lst_cls} which is an ISCAM
model list
}
\description{
Translate the gear names and abbreviations in the models list for all
models to either French of English
}
\details{
Depends on the return value of fr() which is just the value
of the R global option \code{french}
}
