% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-bio-depl.R
\name{table_bio_depl}
\alias{table_bio_depl}
\title{Create a table containing biomass and depletion estimates}
\usage{
table_bio_depl(
  model,
  syr,
  digits = 3,
  xcaption = "default",
  xlabel = "default",
  font.size = 9,
  space.size = 10,
  placement = "H",
  tabular.environment = "tabular",
  translate = FALSE
)
}
\arguments{
\item{model}{An iSCAM model object as created in \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{syr}{Start year}

\item{digits}{Number of decimal places in the table}

\item{xcaption}{Table caption}

\item{xlabel}{Latex table label}

\item{font.size}{Size of font}

\item{space.size}{Size of spaces between cells}

\item{placement}{Latex table placement string, eg. "H" or "tbp"}

\item{tabular.environment}{See \code{\link[xtable:print.xtable]{xtable::print.xtable()}}}

\item{translate}{Logical. If \code{TRUE}, translate to French}
}
\value{
an xtable
}
\description{
Make a table with both spawning biomass and depletion in it.
Based on \code{\link[=table_ts_values_mcmc]{table_ts_values_mcmc()}}, but wider with both and extra headers
}
