% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of catch by gear type}
\usage{
table_catch(
  catch_df,
  start_yr = NULL,
  digits = 2,
  gear_rm = available_gears,
  gear_col_widths = "5em",
  ...
)
}
\arguments{
\item{catch_df}{A data frame as output by \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}

\item{start_yr}{The year to start the table or \code{NULL} to include all data}

\item{digits}{Number of decimal places for the values in the table}

\item{gear_rm}{A vector of gear types to remove from the table. Look at the
function code to see the list of available gears (\code{available_gears})}

\item{gear_col_widths}{Width to make the columns for the gear types}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of catch by gear type
}
