% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-catch.R
\name{table_catch}
\alias{table_catch}
\title{Create a table of catch, by Landings and Discards}
\usage{
table_catch(
  catch_df,
  start_yr = NULL,
  by_area = FALSE,
  gear_col_widths = "5em",
  scale_factor = 1000,
  ret_df = FALSE,
  bold_header = TRUE,
  ...
)
}
\arguments{
\item{catch_df}{A data frame as output by \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}

\item{start_yr}{The year to start the table or \code{NULL} to include all data}

\item{by_area}{Logical. If TRUE, make a table by area (grouped as 3CD and
5ABCDE)}

\item{gear_col_widths}{Width to make the columns for the gear types}

\item{scale_factor}{Value to divide all the catches by to make the table
in tonnes}

\item{ret_df}{If \code{TRUE} return a data frame, if \code{FALSE}, return an
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of catch, by Landings and Discards
}
