% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-decisions.R
\name{table_decisions}
\alias{table_decisions}
\title{Create a decision table for iSCAM models}
\usage{
table_decisions(
  model,
  bo_refpts = c(0.2, 0.4),
  bmsy_refpts = c(0.4, 0.8),
  bold_header = FALSE,
  digits = 2,
  ret_df = FALSE,
  col_widths = NULL,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{bo_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.2B0 and 0.4B0 by default}

\item{bmsy_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.4BMSY and 0.8BMSY by default}

\item{digits}{Number of decimal places for the values in the table}

\item{ret_df}{Logical. If \code{TRUE} return a data frame with the values,
instead of the \code{\link[knitr:kable]{knitr::kable()}} formatted table}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Produce a decision table for the given iscam model
}
