% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-growth-params.R
\name{table_growth_params}
\alias{table_growth_params}
\title{Create table of the growth parameters which were input into the ISCAM model}
\usage{
table_growth_params(
  model,
  col_widths = NULL,
  ret_df = FALSE,
  digits = 2,
  alpha_digits = 7,
  ...
)
}
\arguments{
\item{model}{An iscam model object (class \link{mdl_cls})}

\item{col_widths}{Widths for columns, except the Parameter column
the \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{ret_df}{Logical. If \code{TRUE}, return the \link{data.frame} and not}

\item{digits}{The number of decimal places to report on all except the
\code{alpha} parameter}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{alpha_digitws}{The number of decimal places to report on the \code{alpha}
parameter}
}
\value{
Either a \link{data.frame} or a \code{\link[csasdown:csas_table]{csasdown::csas_table()}}, depending on
the value of \code{return_df}
}
\description{
Create table of the growth parameters which were input into the ISCAM model
}
\details{
iSCAM takes these values as inputs in the DAT file and creates a vector
of maturities-at-age by calling \code{plogis(a50, sd50)} where \code{a50} and \code{sd50}
are the outputs of this function, \code{sd50} is actually the shape parameter,
not a true standard deviation of values.
}
