% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-param-phi-mpd.R
\name{table_param_phi_mpd}
\alias{table_param_phi_mpd}
\title{Create a table comparing phi parameter}
\usage{
table_param_phi_mpd(model, digits = 2, french = FALSE, col_widths = "5em", ...)
}
\arguments{
\item{model}{An iSCAM model as output from \code{\link[=load_iscam_files]{load_iscam_files()}}}

\item{digits}{Number of digits to show}

\item{french}{If \code{TRUE} translate to French}

\item{col_widths}{Widths for columns, except the Parameter column}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
An \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of comparisons of estimates of the phi parameter for the
Dirichlet Multinomial for MPD models.
}
