% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ref-points-mcmc.R
\name{table_ref_points_mcmc}
\alias{table_ref_points_mcmc}
\title{Make a table of reference points for MCMC iSCAM models}
\usage{
table_ref_points_mcmc(
  models,
  type = c("median", "ci"),
  bo_refpts = c(0.2, 0.4),
  bmsy_refpts = c(0.4, 0.8),
  digits = 2,
  probs = c(0.025, 0.5, 0.975),
  model_col_widths = NULL,
  ord = NULL,
  ...
)
}
\arguments{
\item{bo_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.2B0 and 0.4B0 by default}

\item{bmsy_refpts}{Vector of two proportional values for the limit reference
point and Upper stock reference. Values are 0.4BMSY and 0.8BMSY by default}

\item{digits}{Number of decimal places for the values in the table}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{ord}{A vector of numbers signifying the order in which the rows should
appear in the table based on what they were by default. If \code{NULL} the
default order will be shown, then based on what you see, you decide what
row order to supply to this argument. If too short or too long a vector
is supplied, an error will be thrown}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of reference points for MCMC iSCAM models
}
