% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-sens-params.R
\name{table_sens_param_changes}
\alias{table_sens_param_changes}
\title{Create a table summarizing changes for sensitivity models}
\usage{
table_sens_param_changes(
  sens_desc_vec,
  sens_change_vec,
  ret_df = FALSE,
  col_widths = NULL,
  bold_header = TRUE,
  ...
)
}
\arguments{
\item{sens_desc_vec}{A vector of sensitivity descriptions}

\item{sens_change_vec}{A vector of sensitivity changes}

\item{ret_df}{Logical. If \code{TRUE}, return a data frame instead of the table}

\item{col_widths}{A vector of column widths. See \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{...}{Arguments passed to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table summarizing changes for sensitivity models
}
