% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-ts-values.R
\name{table_ts_values_mcmc}
\alias{table_ts_values_mcmc}
\title{Create a table of time series estimates for MCMC iSCAM models}
\usage{
table_ts_values_mcmc(
  models,
  model_desc = NULL,
  value = c("sbt", "rt", "ft", "ut", "depl"),
  start_yr = NULL,
  digits = 2,
  probs = c(0.025, 0.5, 0.975),
  model_col_widths = NULL,
  bold_headers = TRUE,
  ...
)
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}

\item{model_desc}{A description for the models to be shown in multicolumn
headings in the table (if more than one model only). If \code{NULL}, names will
be created (Model 1, Model 2, etc.)}

\item{value}{An output value to produce the table for. Can be
'sbt' (spawning biomass), 'rt' (recruitment), 'ft' (fishing mortality),
'ut' (exploitation rate), or 'depl' (depletion)}

\item{start_yr}{Year to start the table with. If \code{NULL}, all years will
be included}

\item{digits}{Number of decimal places for the values in the table}

\item{probs}{A 3-element vector of probabilities that appear in the output
data frames. This is provided in case the data frames have more than three
different quantile levels}

\item{model_col_widths}{Widths for columns, except the Parameter column}

\item{bold_headers}{If \code{TRUE}, make all column headers bold}

\item{...}{Arguments to pass to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Make a table of medians and/or credible intervals for values such as
biomass, recruitment, depletion, and fishing mortality
}
